+++ .gitignore
@@ -1,0 +1,4 @@
+*.jl.*.cov
+*.jl.cov
+*.jl.mem
+/Manifest.toml
