+++ README.md
@@ -1,0 +1,4 @@
+# rxciphers
+
+[![Build Status](https://github.com/rxwop/rxciphers.jl/actions/workflows/CI.yml/badge.svg?branch=master)](https://github.com/rxwop/rxciphers.jl/actions/workflows/CI.yml?query=branch%3Amaster)
+[![Coverage](https://codecov.io/gh/rxwop/rxciphers.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/rxwop/rxciphers.jl)
