--- fitnesses.csv
@@ -1,300 +1,0 @@
--0.9978997318372116
--0.7669071536024257
-0.0
--0.29813521672371746
--0.20766610887350012
--0.24102592528043054
--0.7668908497062983
--0.28869578068676116
--0.1716859573308076
--0.7996261666171973
--0.5480272858079136
--0.11460314440026931
--0.5814464999602382
--0.17119401785341637
--0.4000845160359745
--0.5855284070448228
--0.6283363553681411
--0.3129751789047335
--0.5235542532353854
--1.3125329015897513
--0.9900231834087077
--0.36904434766846084
--1.4162406570004258
--1.356383223451572
--0.7926105590099342
--1.3857837036900005
--1.5788427050410991
--0.9225313818595735
--0.49615859105260807
--0.6992470319520079
--1.5281426106584242
--1.0734382456310083
--1.9170148704148424
--1.9385701484867166
--1.2929024268286815
--1.2946858401443961
--1.3913180286094544
--0.7788892915424617
--1.1255337550721016
--1.1117811752782254
--0.9602238703970407
--2.2748191431920484
--1.6096404484520361
--1.8169492766658015
--0.950754044794226
--1.4893432020394526
--1.554859684088175
--1.8032151124431604
--1.773420923802429
--1.8270637376167738
--2.5663906114175195
--2.573778753399586
--1.7202426480086084
--1.3466033857252642
--1.3341467533336537
--1.573120091421206
--2.2168081217493496
--1.7061307054397767
--2.767466427104801
--0.6311892850109322
--2.719521649432176
--2.465061334831331
--1.3735493888426653
--2.6514828554372363
--1.0282606583079197
--1.8449291323714867
--0.9128976101685602
--1.7903324446337328
--1.798216087867413
--1.9765348577509956
--1.2374789504549506
--1.7706669237864459
--1.8825921861342847
--1.7265799953078957
--1.043297702514142
--3.053810545845125
--1.5750872430345613
--2.6429123063437068
--2.0789943406626525
--0.9346811907793287
--2.32978806389747
--2.1083751378257416
--2.46537560450379
--2.85661239884082
--2.6366909954904045
--2.370192769048476
--1.7471216617446546
--1.5078922807467539
--2.0484959317074214
--1.800754734930849
--2.495418519124991
--2.0509746517000167
--1.6389534694769026
--1.4396976010891551
--2.2115270734042287
--3.6264759934060002
--2.532925868195684
--2.5652473271360403
--2.6762283759620358
--2.4026842611513306
--3.1545817430184835
--1.6193167567114601
--2.7101416232982016
--2.0677544936982413
--2.612271506165457
--2.9320902914688842
--3.268583399431683
--2.689788547315705
--1.7371189671342782
--2.4163480064375475
--2.65887320206562
--3.520721330890324
--3.4764379395540095
--2.5585044262561585
--2.533624808346473
--2.5889962520290224
--2.9908680414207076
--2.243070311801895
--2.3453881554160523
--2.205060099047828
--1.8582865185970618
--2.085671341589606
--2.7220552477249695
--3.0104507344503446
--2.7136697546507698
--2.924199245570179
--2.9937588609555537
--2.9752469024339225
--2.552300351448814
--2.6182850980743977
--2.132609475334019
--2.189883899801041
--3.2694901660993496
--2.609396500339259
--1.994356173506473
--2.6360450284445767
--3.0985611069983596
--2.715877454733917
--2.3661601742863576
--3.187758109503843
--2.3712880434877768
--2.7757482742730293
--1.95766514274007
--2.5676574372622225
--1.6852513297386658
--2.4360963372967346
--2.82363379009106
--3.382462403830014
--2.1082574183654845
--2.9598537011846977
--2.904384283526821
--2.1930676910204348
--2.68598417986584
--3.0576407602918465
--2.2993527874468587
--2.8603853937059682
--1.5144744111350903
--2.3602869853975115
--2.5220181472884686
--2.116191370057247
--2.282993067908188
--2.7419452911967044
--1.8232621680666785
--3.3857520027393733
--2.827369047094945
--3.602552333429598
--1.9253455000985618
--2.4420469946804717
--3.9551557582115953
--3.510771287668528
--3.0025784702418443
--4.204881466418439
--4.482172241656413
--3.8074737119627704
--3.7217073682650526
--3.0370155594835326
--2.9503891702847254
--3.6818183293496176
--2.7518805505019293
--3.131259052599134
--3.1148665430296942
--2.073423030429513
--2.572958544022806
--3.1994194908148392
--3.1977289803775637
--2.9171406547927026
--2.8964736110497444
--3.6723464788752693
--2.4075813321218478
--2.811838074778394
--3.655502267552296
--2.452580022107421
--4.020834121508302
--2.805030288278428
--2.1180998847867256
--2.8491112490416484
--3.7969499621031035
--3.4407610229430237
--2.934033317003091
--2.984824161979704
--4.15303525861276
--3.200366397692071
--3.5377329665561392
--2.467399852653802
--3.5724180383217403
--2.3880496724708618
--2.9689981627231115
--3.075308704332029
--2.4268926852778128
--3.019960148722915
--2.625195610765382
--2.8037252727388395
--3.0168610216092713
--2.812336420095976
--3.8962064744250133
--2.0726729676210383
--3.4388643282251703
--4.1127497324158115
--3.2976171890365276
--2.6702772460455844
--2.503090640604782
--2.8548111843793773
--2.8138376862754493
--2.925793729137938
--2.729661074738482
--4.768178402789835
--3.3855209197485516
--4.4056321792972595
--3.27475743222164
--3.893652218155631
--3.906747616480218
--3.510221999487505
--3.4762873913968786
--3.0847122999309047
--3.262504354358441
--3.807443401225708
--3.63646931602712
--4.506017938033449
--3.442168689791729
--4.50830503846852
--3.738023950663707
--3.2266609304899836
--2.4239851859934634
--3.105781591229639
--3.299315459189941
--2.918159856759802
--3.557986074631886
--2.76939511626197
--3.0190635530491123
--2.981186524909628
--3.0543935871009893
--3.156545248410616
--3.98576580587521
--3.829473372923455
--3.0506046024211173
--3.7963625677910064
--3.522484373951043
--3.9026218009936615
--3.6181324358187474
--3.2393329292106063
--3.6989555504982414
--3.7052932834439414
--3.0551208675614383
--3.6243557322334627
--3.552625705542626
--3.1048146853090293
--2.9585370946349796
--3.4589741275649075
--4.4621998900096935
--3.9517502265161726
--2.9513701437336897
--4.293481847430734
--3.8193130588823516
--3.9569105489601304
--3.4358893195805367
--3.4911813632677537
--3.310781975591903
--3.5786762379001864
--2.7900317351889923
--3.1812618375970905
--3.6891856561065426
--2.4567022442159168
--3.411351842950568
--3.449475728039401
--4.286996498755562
--2.980375099208861
--3.0520419905155363
--3.5241756451114243
--4.143275935114677
--3.633481632412477
--3.6609224017056876
--3.178772486790079
--3.0055941270214968
--3.2630820059308716
--3.313840890411247
--3.3250758768347275
--3.4048112604540757
--3.6964198078001154
--3.5365424360312554
--3.293925468243943
