--- information_densities.csv
@@ -1,300 +1,0 @@
-0.02734569163136058
-0.03049877483328345
-0.038830658017577235
-0.033630680290619885
-0.03456848506511889
-0.034882891977670916
-0.029398523548191946
-0.03388753576446746
-0.035426157270514125
-0.030277657562163718
-0.033380493533847755
-0.03567909224662724
-0.03189920378764022
-0.035357274421611073
-0.03400276418989634
-0.02948368073216656
-0.029228462879187234
-0.03275136598787931
-0.029589967585023302
-0.025639820536445868
-0.025595753122062526
-0.03144031973376826
-0.023596916848048544
-0.022201493581928133
-0.028205024110377508
-0.025666902688302073
-0.0235784907679637
-0.027277863584411755
-0.03127691732311888
-0.02886903521494677
-0.020253220761647484
-0.02290547859990255
-0.020759949449061565
-0.0170562340686482
-0.022670543959157463
-0.019293934036265448
-0.02370499384899638
-0.028459292216573395
-0.024650051995465994
-0.025606074573484072
-0.0274887066969427
-0.019457184947683892
-0.020414029989218818
-0.019134174677492233
-0.0256864726291246
-0.01949449593377839
-0.018693149207012816
-0.020825296861421094
-0.0209252329958165
-0.017495528608956867
-0.01384967824155936
-0.013859978052283672
-0.015636790209873297
-0.02112562683402519
-0.021040525563466694
-0.022450346267759436
-0.017018407445894113
-0.021812573396442048
-0.010831606067146282
-0.027065040094835346
-0.010650170486357524
-0.015129457043456362
-0.01875632600565247
-0.012481376831803281
-0.01987906625537228
-0.01849906064923168
-0.02482690733600368
-0.014444371414785273
-0.017760233464074737
-0.017048387045248248
-0.023127356664661036
-0.015452855431258072
-0.012851950257578793
-0.017882062478339122
-0.022982252115754775
-0.009983075369912164
-0.01580144046565035
-0.009659241448455091
-0.011733704233005682
-0.02443251889023479
-0.01239157878006214
-0.015833696403381836
-0.010995390610844424
-0.008274793839795791
-0.010587307188754558
-0.01390163197083183
-0.015393630189102225
-0.019898798375687754
-0.012812269941645305
-0.016227369058299246
-0.010641292248369183
-0.013477874918248034
-0.015142657594536146
-0.01644942541657546
-0.01386858855602835
-0.004895926553867724
-0.01312007147647165
-0.008990509666305435
-0.00960037956838633
-0.011049896884920862
-0.010624667669700827
-0.019904892076696434
-0.010859942740945705
-0.011623761254476506
-0.01175365516505611
-0.009149980308919226
-0.006389429946888292
-0.008502750759016742
-0.017356909886399485
-0.013537709080981841
-0.009259184876720283
-0.006369084791218782
-0.006405083456560767
-0.007218006906230201
-0.008673237034269476
-0.005725508651678979
-0.009738159776508049
-0.014012483184829172
-0.010340916708106107
-0.01060679579730639
-0.011747872722834174
-0.013835715282328742
-0.0074418240550384596
-0.007710301212650705
-0.007253446283693066
-0.007151284567335381
-0.009148244946165285
-0.007835855363023114
-0.007769171156400745
-0.008896508495263689
-0.01034341431523167
-0.006819743112123966
-0.004576050532395293
-0.006879854201627289
-0.015263859435299842
-0.0071390274046923
-0.00683888081593409
-0.005291734561392758
-0.011289594282187565
-0.006339895766245718
-0.010409229590443658
-0.007620713141410107
-0.013319539835560334
-0.00620918232081813
-0.015283142656445037
-0.008223669976796799
-0.007350395993910297
-0.007615368139797241
-0.01326804390367986
-0.009864302840934869
-0.006097063328243557
-0.01504898260408094
-0.006713592609829178
-0.007174356731916021
-0.00944805644572444
-0.009388246335024775
-0.015183147030116473
-0.009683862195669413
-0.009072265739320962
-0.00920622905135815
-0.010936132297135082
-0.0064094181074851875
-0.015439120018006847
-0.00386501985706289
-0.006556852442080263
-0.004902460953979393
-0.014042866883006155
-0.010428876429710243
-0.003164578815715302
-0.004101640313996846
-0.005655563529308516
-0.0015372172030119202
-0.0029178684683976996
-0.0022077857258233635
-0.0036722075258595055
-0.005404151426898984
-0.008128563876597994
-0.002822981781649858
-0.008470304907714166
-0.005403865783584649
-0.004054579923277402
-0.01151902639630098
-0.009990892977244314
-0.006257764892340312
-0.004521230575881496
-0.007443435637489948
-0.004389324143429185
-0.004878788101715159
-0.006539867706151448
-0.004784288412627902
-0.004431664836702431
-0.005689477379412722
-0.0029881576322763337
-0.005199706581287877
-0.011446177550901382
-0.005241048276837468
-0.0029671735630110163
-0.003908001090891844
-0.005528942086620248
-0.007146846652496208
-0.0014343127607272521
-0.0034729756348328422
-0.003444726481898641
-0.007365903154481934
-0.0035826612868099517
-0.009504371977796744
-0.0034578353446237386
-0.00553935726099386
-0.008866336757660243
-0.002589453814846543
-0.005547860346251099
-0.0033429367683286415
-0.0049456105813829865
-0.00643223428425177
-0.0020991008972019998
-0.010560827123576644
-0.0032658367840587843
-0.0017551536930033862
-0.005980704475565091
-0.007281881920086164
-0.009230100891768345
-0.0066569504537854485
-0.004232373687551139
-0.008312804063589141
-0.005654752088704029
-0.00013026266489113425
-0.004181243008226865
--0.00020905439635932055
-0.003899498543977309
-0.0018572284993103594
-0.0021366701216720685
-0.002655673677956827
-0.003227954697555546
-0.0029547928942683654
-0.0021328140509737636
-0.0020122068204022625
-0.0019087190505437945
-0.0006382405064053742
-0.003461348085517457
-0.001826772645292074
-0.003121010876718344
-0.003574508373995755
-0.006640231568563415
-0.0022820952685682627
-0.0015383745121919645
-0.003952436827940886
-0.0017567730300092066
-0.002827347899664545
-0.0036021379207582527
-0.004918998171683953
-0.0037035485645633415
-0.0027648068049881402
-0.0018293785297593956
-0.0020692126058836635
-0.0038049948700562553
-0.0017213902789539048
-0.0011698409969795875
-0.002582132642023005
-0.0017770904800089582
-0.004460909863428064
-0.0010352826419044447
-0.004037782292249048
-0.007057492111006702
-0.0013661304535173002
-0.00361411026107028
-0.0049222976604283986
-0.00417190355187369
-0.0031871585857405904
--0.00026860767535463737
-0.0028464014940157156
-0.005420452464080405
-0.0014669659432167137
-0.0016915739807839003
-0.002347974612756154
-0.004695327216957423
-0.0044327616908024225
-0.0031479272049841356
-0.002619241146710749
-0.005830882282296557
-0.0038876076006202354
-0.0016367819891032198
-0.0031740968205538067
-0.003928311559951325
-0.0013626266795263575
-0.0003246342464313656
-0.0027688213622480973
-0.00662265980612986
-0.0015728193973858495
-0.0009979958406249878
-0.002741211297196955
-0.0037252245771535186
-0.002088758580481631
-0.004644249290310931
-0.0034404104197740537
-0.0017013915923662637
-0.003058890612172682
-0.0009757542043136169
-0.0017038990644628882
-0.002795805442858068
-0.0019009614110920836
