+++ project.toml
@@ -1,3 +1,20 @@
-name = "CipherEngine"
-uuid = "4ug30569-f782-26z3-1bc8-f53579315bfe"
-authors = ["SquidSinker", "rxwop", "Nayoar"]
+name = "rxciphers"
+uuid = "5286dc34-9e20-4123-8e59-fcde100af23f"
+authors = ["rxwop <robiwop@gmail.com>", "Nayoar"]
+version = "0.1.0"
+
+[deps]
+FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
+IterTools = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
+JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+
+[compat]
+julia = "1"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
