--- src/m10.jl
@@ -1,17 +1,0 @@
-include("cracks.jl")
-include("substitution.jl")
-include("transposition.jl")
-include("grid_shift.jl")
-
-T = Txt("IHRHG FTVEE EENNP OAIAO LARAE IEETI NNBKL SPREE ETNRM RAHSE NEUNI IUMND DTANW REAMO AYDMC DJNER ETSNA DAANF BYOAY MIDOI LJKRB GWHIU ANFTG BRESV MTFTA OAEEO OSERY OEEAS ANDTE TVODI OERSM SNNLC SHEEN NURYA LIAWN WPUDS UANNN EECMR CENAN ROITP LETBR RUYGH STNEN LOLGI OUEEG IASTF OLEEY RSMKI ITTIT DHCFS INSOU CEIIS EANOT EREME SINGI IANEA CVWEN MEIDR AEOTR DSRON RCOTT TTATG CUOLB EMEST DRFAI RIEAA NMTDU ROIEL DDLEE NFCUP AEWHT TLHNG CNRSN ATDRA NTELI ICNFD EEORE IVVEE SONIE HFSNH NATLR EAWUO OSTAT RISOS AOTSD SSASR ISIIE AROND ESTRU OWEAE SNLHT MENEI GKDDS TBELD UIPTN WAELM SDERT IOYEM GHWLO DPUUL VSTTD EMLGP CEODH PITRE AHFRH SYABT UNDOE COEEP RSSTD WIENM NMOID FRMEI NHYEC SSUDN MUNDS OSTAU CMDEN RADVT TAROT SNLAI SNISK UENCA EPELG ATSDE EAIRD SHBIT OEGPB ODEEF DISRO DNIST TSUEU CTSAC GWDRO ENOON IRWDE NNTOD ICINO MNLLO SSEOT FREOL MACSO OPSDO HDFIR GKTNE CIROE ROOHW BCCPN OARDI TNBKC AENOB IEUDT YEYOO FIGAM UDHGR CBONE ORHHI REHEE ROINS DRNUT ETIAA IXUAS OIEIE IDTRT ENNZM CTOLE CSCID NENCN PVOTT DOEGE OOOHD DRNFW UHENN AIURT SEDDD SSIYS RHHEL NEOSO USNEL VTMDG NWDTD AAMEL XANEF EIENR ECTOU NSKKO DTASH NGGEE IOTEE LEUEN RISOR OANGA LNTFC DRVUI NMDDA UHNLN MIVAD FNEBG OARRI TEEOA ERENN REIRE DERDI EDRUL DBITI SYATE NEOLF DSHER NRETB EIILO TOBDT OWRIR NSDAN HEEAU EETNS TNMRE MTSME OTLES TUHEI IIAIA TPGME HTEAA BMENT CNEAT EBNID ERGNT IDCRF NWCRH RNRDP REOOR RATOO USARC NBGTH HMLTE TPIAE EICFE FNLDT OPNNA INIHT AATTN TIDIO EETNR EEMRH EHABH GTTSE NUOEI ESTDE IOSIN HCTST BNNDR CANEN HNAWT NIYTH AITAS PSSCO EAOHA THBDO TBRDT MECIN WEGEG HEREE RNBOL HEKAN SNDFA SRXIY EOAGL TNESE MIAII YNGAX URETL ATEUG ENTPH NCCLS EHTID RLIAE PIBEA IYOOC VNIWR TLYEO AFITB HWOOL TATMP OFSGF LWIOD NFVTA TEDAD REUET HCOEW AMWNE NGNRE WPIAA TDEPO SIETE WMAHE SRRAE HIAST FHATN NERDT CSEUM NSDNE TOTTN EWRIT MLDEH GDMEH EABHS WSSOA SRTDT HDYDE GEOHO EFNIG IAHEI YMEAT UOEST ERXUA CEDVH TTAHN NOFSR TLNBU IERTI AEEDM WTPCR REDEE LHETH IEEEU UEISL IRITM CHYEH LVBTO TVAAF OBDAO PMWLN ODTLT SNHWO ATIBE EDXOT TTEIR THSPM OONGN EPHBO SOUOL FNHAG DTESI GLTNE TTEAT TRDMO LYVSM ESEAG ENSNE EOFNT OWTFP RSRWN RIRFA EGHEF LISOU IOTRA YTNHI HSPEE SFLIE DIGRS EEAIY ARESA HNNPI GESTT AXOSE IIYGA ENGPE TDNGE FEHOT EARPC WENNO NTBRD ERVLO EOTNO IGMRN EEOHR ESSOS DUDRL OFATL EEECI HNTET BDHUC RHBAO EASIT WTNSS NMHOO LLAON IGOHC ETSOS ECDRI ONIHX AEMSA ROSRU NRCAO HHOIV AIAAT NISVT HSNOH OSWUE TYISE REYOH NTEER LRNIM SLEHG TNTEE OROST OCENE NAFEL EBHDT OWHEI DOSER TTSAO XTUIE RSTFS TINPW IDIEE ELWCW AIFHC COAHL OHCVP NERXO HHACE EIEEE CEAOA NWTMM TCODN HDEED BFENA SCEAO SBTMK LNSGO HCNBP UDRKY OEYSE DETTS RENFH HEHET SOSUW UADEE STAOO EYTBA HLXWS YNIML EFATE TIAOS RPAIE ATNNI SROHS HMNBN AGROE BTOSE NEHHE SELDO MBOSS FTUED UEHAE STUOA XGTSD ITSET PTDBW DFRTA IYRTR EETIE RESND HOALS TOCIE TRNHT EDCPM EOEST HOOAE EARDA ERUYN SNHHE SREEF OOLAO TTNTT SIOWR ILOSE TNEIA ERTPN FHYOP EBRNR ANOIR NAMOD HUTMT ESNST HUBKP OMNEI CITSC REVAE FIKQB HCEIC CENHT ASTER SUAOS OOIAG FFNST POEEE IIEPE NOSRM HTTOE MRAAD ATUNI SEOUS MMALN KBSBT EFCEU FNRDU NYASO TESCE RGOBO TYGND HUNHE AITEU EEART NIOUY PEORW UNTNE IEAOC GRTCR RAMNN TSCIU QSHTI IAGTN HREEI ESKAU DAAXT SHUTE HAHER ESOEI ECOVT BOSGD IJTNN NTOUM SEALU IAFAO DFNAE INDAF TYTAT ENKGW RNOAE TRYES OSTND ATWNR SNRUN ESEOE ISNEE OITMR MNWTT OIAOT TRNNW ATNWT YWSYT IINOR DAAOO HAEHE SINWJ GTRHS NVNRT TEEON GOCEH KEEFR RISKH ESEOG KHIRS IFPAG DNIFO ADOEA REWKG HTITO WREEA ERCOR UWFUT OHEUE CFTNC RENOT OEOIT GACET RAOYO ICEIU TNEES ITDIO DHEOE NLSHO EISUS BDTNA RUOTE NERSD EARST DTADE EPEWT IUDAN ITPRI PHNIO ENOET EPTRV GAGRM RNUHU IRRVS EISGC LICMA USSTI OOMAN TEAEO YSDFR REEDN TDSRE OANSA YFGFR IMCTO LOAFM YITEM RIUTL SPTBE GFNNP NTSAU AOKGT YAVEI OMNYT RNOEM HIGSS KANPH FCRNR TATOO EATPE DOUHT NOEWT BELDS EDIDD EBOCB EEPTI OVEAE ATSRS EEEID LPIWR SPUHB ADVEH TNNFL ROENT OAIHU SEHNW EFFGD LTTOS EETEU DDHEV REEUN FUETA AWHAE OVTMC ARRHI TSWLH ONHEU")
-tokenise!(T)
-
-b = blocks(T, 130)
-
-shifts = [18, 15, 12, 11, 4] .+ [4, 3, 3, 3, 4]
-
-C = Encryption(GridShift( - shifts, 26), Permutation([4,2,3,1,5]))
-
-for i in b
-    println(untokenise(C(i)))
-end
