+++ src/recycling/n-m10.jl
@@ -1,0 +1,94 @@
+include("charspace.jl")
+include("tuco.jl")
+include("periodic_substitution.jl")
+include("cracks.jl")
+include("transposition.jl")
+using Combinatorics
+
+t = "IHRHG FTVEE EENNP OAIAO LARAE IEETI NNBKL SPREE ETNRM RAHSE NEUNI IUMND DTANW REAMO AYDMC DJNER ETSNA DAANF BYOAY MIDOI LJKRB GWHIU ANFTG BRESV MTFTA OAEEO OSERY OEEAS ANDTE TVODI OERSM SNNLC SHEEN NURYA LIAWN WPUDS UANNN EECMR CENAN ROITP LETBR RUYGH STNEN LOLGI OUEEG IASTF OLEEY RSMKI ITTIT DHCFS INSOU CEIIS EANOT EREME SINGI IANEA CVWEN MEIDR AEOTR DSRON RCOTT TTATG CUOLB EMEST DRFAI RIEAA NMTDU ROIEL DDLEE NFCUP AEWHT TLHNG CNRSN ATDRA NTELI ICNFD EEORE IVVEE SONIE HFSNH NATLR EAWUO OSTAT RISOS AOTSD SSASR ISIIE AROND ESTRU OWEAE SNLHT MENEI GKDDS TBELD UIPTN WAELM SDERT IOYEM GHWLO DPUUL VSTTD EMLGP CEODH PITRE AHFRH SYABT UNDOE COEEP RSSTD WIENM NMOID FRMEI NHYEC SSUDN MUNDS OSTAU CMDEN RADVT TAROT SNLAI SNISK UENCA EPELG ATSDE EAIRD SHBIT OEGPB ODEEF DISRO DNIST TSUEU CTSAC GWDRO ENOON IRWDE NNTOD ICINO MNLLO SSEOT FREOL MACSO OPSDO HDFIR GKTNE CIROE ROOHW BCCPN OARDI TNBKC AENOB IEUDT YEYOO FIGAM UDHGR CBONE ORHHI REHEE ROINS DRNUT ETIAA IXUAS OIEIE IDTRT ENNZM CTOLE CSCID NENCN PVOTT DOEGE OOOHD DRNFW UHENN AIURT SEDDD SSIYS RHHEL NEOSO USNEL VTMDG NWDTD AAMEL XANEF EIENR ECTOU NSKKO DTASH NGGEE IOTEE LEUEN RISOR OANGA LNTFC DRVUI NMDDA UHNLN MIVAD FNEBG OARRI TEEOA ERENN REIRE DERDI EDRUL DBITI SYATE NEOLF DSHER NRETB EIILO TOBDT OWRIR NSDAN HEEAU EETNS TNMRE MTSME OTLES TUHEI IIAIA TPGME HTEAA BMENT CNEAT EBNID ERGNT IDCRF NWCRH RNRDP REOOR RATOO USARC NBGTH HMLTE TPIAE EICFE FNLDT OPNNA INIHT AATTN TIDIO EETNR EEMRH EHABH GTTSE NUOEI ESTDE IOSIN HCTST BNNDR CANEN HNAWT NIYTH AITAS PSSCO EAOHA THBDO TBRDT MECIN WEGEG HEREE RNBOL HEKAN SNDFA SRXIY EOAGL TNESE MIAII YNGAX URETL ATEUG ENTPH NCCLS EHTID RLIAE PIBEA IYOOC VNIWR TLYEO AFITB HWOOL TATMP OFSGF LWIOD NFVTA TEDAD REUET HCOEW AMWNE NGNRE WPIAA TDEPO SIETE WMAHE SRRAE HIAST FHATN NERDT CSEUM NSDNE TOTTN EWRIT MLDEH GDMEH EABHS WSSOA SRTDT HDYDE GEOHO EFNIG IAHEI YMEAT UOEST ERXUA CEDVH TTAHN NOFSR TLNBU IERTI AEEDM WTPCR REDEE LHETH IEEEU UEISL IRITM CHYEH LVBTO TVAAF OBDAO PMWLN ODTLT SNHWO ATIBE EDXOT TTEIR THSPM OONGN EPHBO SOUOL FNHAG DTESI GLTNE TTEAT TRDMO LYVSM ESEAG ENSNE EOFNT OWTFP RSRWN RIRFA EGHEF LISOU IOTRA YTNHI HSPEE SFLIE DIGRS EEAIY ARESA HNNPI GESTT AXOSE IIYGA ENGPE TDNGE FEHOT EARPC WENNO NTBRD ERVLO EOTNO IGMRN EEOHR ESSOS DUDRL OFATL EEECI HNTET BDHUC RHBAO EASIT WTNSS NMHOO LLAON IGOHC ETSOS ECDRI ONIHX AEMSA ROSRU NRCAO HHOIV AIAAT NISVT HSNOH OSWUE TYISE REYOH NTEER LRNIM SLEHG TNTEE OROST OCENE NAFEL EBHDT OWHEI DOSER TTSAO XTUIE RSTFS TINPW IDIEE ELWCW AIFHC COAHL OHCVP NERXO HHACE EIEEE CEAOA NWTMM TCODN HDEED BFENA SCEAO SBTMK LNSGO HCNBP UDRKY OEYSE DETTS RENFH HEHET SOSUW UADEE STAOO EYTBA HLXWS YNIML EFATE TIAOS RPAIE ATNNI SROHS HMNBN AGROE BTOSE NEHHE SELDO MBOSS FTUED UEHAE STUOA XGTSD ITSET PTDBW DFRTA IYRTR EETIE RESND HOALS TOCIE TRNHT EDCPM EOEST HOOAE EARDA ERUYN SNHHE SREEF OOLAO TTNTT SIOWR ILOSE TNEIA ERTPN FHYOP EBRNR ANOIR NAMOD HUTMT ESNST HUBKP OMNEI CITSC REVAE FIKQB HCEIC CENHT ASTER SUAOS OOIAG FFNST POEEE IIEPE NOSRM HTTOE MRAAD ATUNI SEOUS MMALN KBSBT EFCEU FNRDU NYASO TESCE RGOBO TYGND HUNHE AITEU EEART NIOUY PEORW UNTNE IEAOC GRTCR RAMNN TSCIU QSHTI IAGTN HREEI ESKAU DAAXT SHUTE HAHER ESOEI ECOVT BOSGD IJTNN NTOUM SEALU IAFAO DFNAE INDAF TYTAT ENKGW RNOAE TRYES OSTND ATWNR SNRUN ESEOE ISNEE OITMR MNWTT OIAOT TRNNW ATNWT YWSYT IINOR DAAOO HAEHE SINWJ GTRHS NVNRT TEEON GOCEH KEEFR RISKH ESEOG KHIRS IFPAG DNIFO ADOEA REWKG HTITO WREEA ERCOR UWFUT OHEUE CFTNC RENOT OEOIT GACET RAOYO ICEIU TNEES ITDIO DHEOE NLSHO EISUS BDTNA RUOTE NERSD EARST DTADE EPEWT IUDAN ITPRI PHNIO ENOET EPTRV GAGRM RNUHU IRRVS EISGC LICMA USSTI OOMAN TEAEO YSDFR REEDN TDSRE OANSA YFGFR IMCTO LOAFM YITEM RIUTL SPTBE GFNNP NTSAU AOKGT YAVEI OMNYT RNOEM HIGSS KANPH FCRNR TATOO EATPE DOUHT NOEWT BELDS EDIDD EBOCB EEPTI OVEAE ATSRS EEEID LPIWR SPUHB ADVEH TNNFL ROENT OAIHU SEHNW EFFGD LTTOS EETEU DDHEV REEUN FUETA AWHAE OVTMC ARRHI TSWLH ONHEU"
+t = replace(t, ' ' => "")
+T = Txt(t)
+
+tokenise!(T, Alphabet)
+U = deepcopy(T)
+# v = copy(U.tokenised)
+
+# max_rails = 10
+# R = []
+
+# for i in 2:max_rails
+#     for j in 0:2i-1
+#         for k in 2:max_rails
+#             for l in 0:2k-1
+#                 r = Railfence(i, j)
+#                 s = Railfence(k, l)
+#                 invert!(r)
+#                 invert!(s)
+#                 push!(R, (r, s))
+#             end
+#         end
+#     end
+# end
+
+# max_fit = -Inf
+# max_key = nothing
+
+# null_Txt = Txt("")
+
+# for (i, (r, s)) in enumerate(R)
+#     global max_fit, max_key, U, v
+#     v = apply(r, apply(s, v; safety_checks = null_Txt); safety_checks = null_Txt)
+#     filter!(!=(0), v)
+#     U.tokenised = v
+#     fit = quadgramlog(U)
+#     if fit > max_fit
+#         max_fit = fit
+#         max_key = (r, s)
+#     end
+#     i % 100 == 0 ? println(i) : 0
+# end
+
+
+v = T.tokenised
+
+h = 25
+w = 5
+
+chunks = [reshape(v[i-(w*h-1) : i], (h,5)) for i in w*h:w*h:length(v)]
+
+function shift_chunk(chunk, shifts)#, perm)
+    out = copy(chunk)
+    for i in 1:w
+        out[i, :] = circshift(out[i, :], shifts[i])
+    end
+    return out#[perm, :]
+end
+        
+shift_chunks(chunks, shifts) = [shift_chunk(chunk, shifts) for chunk in chunks]
+death(chunks, shifts) = vcat(vec.(shift_chunks(chunks, shifts))...)
+
+max_fit = -Inf
+max_key = nothing
+
+# perms = collect(permutations(1:h))
+
+# for (i, p) in enumerate(perms)
+for a in 1:h
+    for b in 1:h
+        for c in 1:h
+            for d in 1:h
+                for e in 1:h
+                    global max_fit, max_key
+                    U.tokenised = death(chunks, [a, b, c, d, e])
+                    fit = quadgramlog(U)
+                    if fit > max_fit
+                        max_fit = fit
+                        max_key = [a, b, c, d, e]
+                    end
+                end
+            end
+            println(a, " ", b, " ", c)
+        end
+    end
+end
+# end
