+++ src/recycling/test_results.jl
@@ -1,0 +1,176 @@
+const orwell_tokens = [9, 20, 23, 1, 19, 1, 2, 18, 9, 7, 8, 20, 3, 15, 12, 4, 4, 1, 25, 9, 14, 1, 16, 18, 9, 12, 1, 14, 4, 20, 8, 5, 3, 
+12, 15, 3, 11, 19, 23, 5, 18, 5, 19, 20, 18, 9, 11, 9, 14, 7, 20, 8, 9, 18, 20, 5, 5, 14, 23, 9, 14, 19, 20, 15, 14, 19, 13, 9, 20, 8, 8, 9, 19, 3, 8, 9, 14, 14, 21, 26, 26, 12, 5, 4, 9, 14, 20, 15, 8, 9, 19, 2, 18, 5, 1, 19, 20, 9, 14, 1, 14, 5, 6, 6, 15, 18, 20, 20, 15, 5, 19, 3, 1, 16, 5, 20, 8, 5, 22, 9, 12, 5, 23, 9, 14, 4, 19, 12, 9, 
+16, 16, 5, 4, 17, 21, 9, 3, 11, 12, 25, 20, 8, 18, 15, 21, 7, 8, 20, 8, 5, 7, 12, 1, 19, 19, 4, 15, 15, 18, 19, 15, 6, 22, 9, 3, 20, 15, 18, 25, 13, 1, 14, 19, 9, 15, 14, 19, 20, 8, 15, 21, 7, 8, 14, 15, 20, 17, 21, 9, 3, 11, 12, 25, 5, 14, 15, 21, 7, 8, 20, 15, 16, 18, 5, 22, 5, 14, 20, 1, 19, 23, 9, 18, 12, 15, 6, 7, 18, 9, 20, 20, 25, 4, 21, 19, 20, 6, 18, 15, 13, 5, 14, 20, 5, 18, 9, 14, 7, 1, 12, 15, 14, 7, 23, 9, 20, 8, 8, 9, 13, 20, 8, 5, 8, 1, 
+12, 12, 23, 1, 25, 19, 13, 5, 12, 20, 15, 6, 2, 15, 9, 12, 5, 4, 3, 1, 2, 2, 1, 7, 5, 1, 14, 4, 15, 12, 4, 18, 1, 
+7, 13, 1, 20, 19, 1, 20, 15, 14, 5, 5, 14, 4, 15, 6, 9, 20, 1, 3, 15, 12, 15, 21, 18, 5, 4, 16, 15, 19, 20, 5, 18, 20, 15, 15, 12, 1, 18, 7, 5, 6, 15, 18, 9, 14, 4, 15, 15, 18, 4, 9, 19, 16, 12, 1, 25, 8, 1, 4, 2, 5, 5, 14, 20, 
+1, 3, 11, 5, 4, 20, 15, 20, 8, 5, 23, 1, 12, 12, 9, 20, 4, 5, 16, 9, 3, 20, 5, 4, 19, 9, 13, 16, 12, 25, 1, 14, 5, 14, 15, 18, 13, 15, 21, 19, 6, 1, 3, 5, 13, 15, 18, 5, 20, 8, 1, 14, 1, 13, 5, 20, 18, 5, 23, 9, 4, 5, 20, 8, 5, 
+6, 1, 3, 5, 15, 6, 1, 13, 1, 14, 15, 6, 1, 2, 15, 21, 20, 6, 15, 18, 20, 25, 6, 9, 22, 5, 23, 9, 20, 8, 1, 8, 5, 1, 22, 25, 2, 12, 1, 3, 11, 13, 15, 21, 19, 20, 1, 3, 8, 5, 1, 14, 4, 18, 21, 7, 7, 5, 4, 12, 25, 8, 1, 14, 4, 19, 
+15, 13, 5, 6, 5, 1, 20, 21, 18, 5, 19, 23, 9, 14, 19, 20, 15, 14, 13, 1, 4, 5, 6, 15, 18, 20, 8, 5, 19, 20, 1, 9, 
+18, 19, 9, 20, 23, 1, 19, 14, 15, 21, 19, 5, 20, 18, 25, 9, 14, 7, 20, 8, 5, 12, 9, 6, 20, 5, 22, 5, 14, 1, 20, 20, 8, 5, 2, 5, 19, 20, 15, 6, 20, 9, 13, 5, 19, 9, 20, 23, 1, 19, 19, 5, 12, 4, 15, 13, 23, 15, 18, 11, 9, 14, 7, 1, 14, 4, 1, 20, 16, 18, 5, 19, 5, 14, 20, 20, 8, 5, 5, 12, 5, 3, 20, 18, 9, 3, 3, 21, 18, 18, 5, 14, 20, 23, 1, 19, 3, 21, 20, 15, 6, 6, 4, 21, 18, 9, 14, 7, 4, 1, 25, 12, 9, 7, 8, 20, 8, 15, 21, 18, 19, 9, 20, 23, 1, 19, 16, 1, 18, 20, 15, 6, 20, 8, 5, 5, 3, 15, 14, 15, 13, 25, 4, 18, 9, 22, 5, 9, 14, 16, 18, 5, 16, 1, 18, 1, 20, 9, 15, 14, 6, 15, 18, 8, 1, 20, 5, 23, 5, 5, 11, 20, 8, 5, 6, 12, 1, 20, 23, 1, 19, 19, 5, 22, 5, 14, 6, 12, 9, 7, 8, 20, 19, 21, 16, 1, 14, 4, 23, 9, 14, 19, 20, 15, 14, 23, 8, 15, 23, 1, 19, 20, 8, 9, 18, 20, 25, 14, 9, 14, 5, 1, 14, 4, 8, 1, 4, 1, 22, 1, 18, 9, 3, 15, 19, 5, 21, 12, 3, 5, 18, 1, 2, 15, 22, 5, 8, 9, 19, 18, 9, 7, 8, 20, 1, 14, 11, 12, 5, 23, 5, 14, 20, 19, 12, 15, 23, 12, 25, 18, 5, 19, 20, 9, 14, 7, 19, 5, 22, 5, 18, 1, 12, 20, 9, 13, 5, 19, 15, 14, 20, 8, 5, 23, 1, 25, 15, 14, 5, 1, 3, 8, 12, 1, 14, 4, 9, 14, 7, 15, 16, 16, 15, 19, 9, 20, 5, 20, 8, 5, 
+12, 9, 6, 20, 19, 8, 1, 6, 20, 20, 8, 5, 16, 15, 19, 20, 5, 18, 23, 9, 20, 8, 20, 8, 5, 5, 14, 15, 18, 13, 15, 21, 19, 6, 1, 3, 5, 7, 1, 26, 5, 4, 6, 18, 15, 13, 20, 8, 5, 23, 1, 12, 12, 9, 20, 23, 1, 19, 15, 14, 5, 15, 6, 20, 8, 15, 19, 5, 16, 9, 3, 20, 21, 18, 5, 19, 23, 8, 9, 3, 8, 1, 18, 5, 19, 15, 3, 15, 14, 20, 18, 9, 22, 5, 4, 20, 8, 1, 20, 20, 8, 5, 5, 25, 5, 19, 6, 15, 12, 12, 15, 23, 25, 15, 21, 1, 2, 15, 21, 20, 23, 8, 5, 14, 25, 15, 21, 13, 15, 22, 5, 2, 9, 7, 2, 18, 15, 20, 8, 5, 18, 9, 19, 23, 1, 20, 3, 8, 9, 14, 7, 25, 15, 21, 20, 8, 5, 3, 1, 16, 20, 9, 15, 14, 2, 5, 14, 5, 1, 20, 8, 9, 20, 18, 1, 14, 9, 14, 19, 9, 4, 5, 20, 8, 5, 6, 12, 1, 20, 1, 6, 18, 21, 9, 20, 25, 22, 15, 9, 3, 5, 23, 1, 19, 18, 5, 1, 4, 9, 14, 7, 15, 21, 20, 1, 12, 9, 19, 20, 15, 6, 6, 9, 7, 21, 18, 
+5, 19, 23, 8, 9, 3, 8, 8, 1, 4, 19, 15, 13, 5, 20, 8, 9, 14, 7, 20, 15, 4, 15, 23, 9, 20, 8, 20, 8, 5, 16, 18, 15, 4, 21, 3, 20, 9, 15, 14, 15, 6, 16, 9, 7, 9, 18, 15, 14, 20, 8, 5, 22, 15, 9, 3, 5, 3, 1, 13, 5, 6, 18, 15, 13, 1, 14, 15, 2, 12, 15, 14, 7, 13, 5, 20, 1, 12, 16, 12, 1, 17, 21, 5, 12, 9, 11, 5, 1, 4, 21, 12, 12, 5, 4, 13, 9, 18, 18, 15, 18, 23, 8, 9, 3, 8, 6, 15, 18, 13, 5, 4, 16, 1, 18, 20, 15, 6, 20, 8, 5, 19, 21, 18, 6, 1, 3, 5, 15, 6, 20, 8, 5, 18, 9, 7, 8, 20, 8, 1, 14, 4, 23, 1, 12, 12, 23, 9, 14, 19, 20, 15, 14, 20, 21, 18, 14, 5, 4, 1, 19, 23, 9, 20, 3, 8, 1, 14, 4, 20, 8, 5, 22, 15, 9, 3, 5, 19, 1, 14, 11, 19, 15, 13, 5, 23, 8, 1, 20, 20, 8, 15, 21, 7, 
+8, 20, 8, 5, 23, 15, 18, 4, 19, 23, 5, 18, 5, 19, 20, 9, 12, 12, 4, 9, 19, 20, 9, 14, 7, 21, 9, 19, 8, 1, 2, 12, 5, 20, 8, 5, 9, 14, 19, 20, 18, 21, 13, 5, 14, 20, 20, 8, 5, 20, 5, 12, 5, 19, 3, 18, 5, 5, 14, 9, 20, 23, 1, 19, 3, 1, 12, 12, 5, 4, 3, 15, 21, 12, 4, 2, 5, 4, 9, 13, 13, 5, 4, 2, 21, 20, 20, 8, 5, 18, 5, 23, 1, 19, 14, 15, 23, 
+1, 25, 15, 6, 19, 8, 21, 20, 20, 9, 14, 7, 9, 20, 15, 6, 6, 3, 15, 13, 16, 12, 5, 20, 5, 12, 25, 8, 5, 13, 15, 22, 5, 4, 15, 22, 5, 18, 20, 15, 20, 8, 5, 23, 9, 14, 4, 15, 23, 1, 19, 13, 1, 12, 12, 9, 19, 8, 6, 18, 1, 9, 12, 6, 
+9, 7, 21, 18, 5, 20, 8, 5, 13, 5, 1, 7, 18, 5, 14, 5, 19, 19, 15, 6, 8, 9, 19, 2, 15, 4, 25, 13, 5, 18, 5, 12, 25, 5, 13, 16, 8, 1, 19, 9, 26, 5, 4, 2, 25, 20, 8, 5, 2, 12, 21, 5, 15, 22, 5, 18, 1, 12, 12, 19, 23, 8, 9, 3, 8, 23, 5, 18, 5, 20, 8, 5, 21, 14, 9, 6, 15, 18, 13, 15, 6, 20, 8, 5, 16, 1, 18, 20, 25, 8, 9, 19, 8, 1, 9, 18, 23, 1, 
+19, 22, 5, 18, 25, 6, 1, 9, 18, 8, 9, 19, 6, 1, 3, 5, 14, 1, 20, 21, 18, 1, 12, 12, 25, 19, 1, 14, 7, 21, 9, 14, 5, 8, 9, 19, 19, 11, 9, 14, 18, 15, 21, 7, 8, 5, 14, 5, 4, 2, 25, 3, 15, 1, 18, 19, 5, 19, 15, 1, 16, 1, 14, 4, 2, 
+12, 21, 14, 20, 18, 1, 26, 15, 18, 2, 12, 1, 4, 5, 19, 1, 14, 4, 20, 8, 5, 3, 15, 12, 4, 15, 6, 20, 8, 5, 23, 9, 14, 20, 5, 18, 20, 8, 1, 20, 8, 1, 4, 10, 21, 19, 20, 5, 14, 4, 5, 4, 15, 21, 20, 19, 9, 4, 5, 5, 22, 5, 14, 20, 8, 18, 15, 21, 7, 8, 20, 8, 5, 19, 8, 21, 20, 23, 9, 14, 4, 15, 23, 16, 1, 14, 5, 20, 8, 5, 23, 15, 18, 12, 4, 12, 15, 15, 11, 5, 4, 3, 15, 12, 4, 4, 15, 23, 14, 9, 14, 20, 8, 5, 19, 20, 18, 5, 5, 20, 12, 9, 20, 20, 12, 5, 5, 4, 4, 9, 5, 19, 15, 6, 23, 9, 14, 4, 23, 5, 18, 5, 23, 8, 9, 18, 12, 9, 14, 7, 4, 21, 19, 20, 1, 14, 4, 20, 15, 18, 14, 16, 1, 16, 5, 18, 9, 14, 20, 15, 19, 16, 9, 18, 1, 12, 19, 1, 14, 4, 20, 8, 15, 21, 7, 8, 20, 8, 5, 19, 21, 14, 
+23, 1, 19, 19, 8, 9, 14, 9, 14, 7, 1, 14, 4, 20, 8, 5, 19, 11, 25, 1, 8, 1, 18, 19, 8, 2, 12, 21, 5, 20, 8, 5, 18, 5, 19, 5, 5, 13, 5, 4, 20, 15, 2, 5, 14, 15, 3, 15, 12, 15, 21, 18, 9, 14, 1, 14, 25, 20, 8, 9, 14, 7, 5, 24, 3, 
+5, 16, 20, 20, 8, 5, 16, 15, 19, 20, 5, 18, 19, 20, 8, 1, 20, 23, 5, 18, 5, 16, 12, 1, 19, 20, 5, 18, 5, 4, 5, 22, 5, 18, 25, 23, 8, 5, 18, 5, 20, 8, 5, 2, 12, 1, 3, 11, 13, 15, 21, 19, 20, 1, 3, 8, 9, 15, 4, 6, 1, 3, 5, 7, 1, 26, 5, 4, 4, 15, 23, 14, 6, 18, 15, 13, 5, 22, 5, 18, 25, 3, 15, 13, 13, 1, 14, 4, 9, 14, 7, 3, 15, 18, 14, 5, 18, 
+20, 8, 5, 18, 5, 23, 1, 19, 15, 14, 5, 15, 14, 20, 8, 5, 8, 15, 21, 19, 5, 6, 18, 15, 14, 20, 9, 13, 13, 5, 4, 9, 
+1, 20, 5, 12, 25, 15, 16, 16, 15, 19, 9, 20, 5, 2, 9, 7, 2, 18, 15, 20, 8, 5, 18, 9, 19, 23, 1, 20, 3, 8, 9, 14, 7, 25, 15, 21, 20, 8, 5, 3, 1, 16, 20, 9, 15, 14, 19, 1, 9, 4, 23, 8, 9, 12, 5, 20, 8, 5, 4, 1, 18, 11, 5, 25, 5, 19, 12, 15, 15, 11, 5, 4, 4, 5, 5, 16, 9, 14, 20, 15, 23, 9, 14, 19, 20, 15, 14, 19, 15, 23, 14, 4, 15, 23, 14, 1, 
+20, 19, 20, 18, 5, 5, 20, 12, 5, 22, 5, 12, 1, 14, 15, 20, 8, 5, 18, 16, 15, 19, 20, 5, 18, 20, 15, 18, 14, 1, 20, 15, 14, 5, 3, 15, 18, 14, 5, 18, 6, 12, 1, 16, 16, 5, 4, 6, 9, 20, 6, 21, 12, 12, 25, 9, 14, 20, 8, 5, 23, 9, 14, 4, 1, 12, 20, 5, 18, 14, 1, 20, 5, 12, 25, 3, 15, 22, 5, 18, 9, 14, 7, 1, 14, 4, 21, 14, 3, 15, 22, 5, 18, 9, 14, 7, 20, 8, 5, 19, 9, 14, 7, 12, 5, 23, 15, 18, 4, 9, 14, 7, 19, 15, 3, 9, 14, 20, 8, 5, 6, 1, 18, 4, 9, 19, 20, 1, 14, 3, 5, 1, 8, 5, 12, 9, 3, 15, 16, 20, 5, 18, 19, 11, 9, 13, 13, 5, 4, 4, 15, 23, 14, 2, 5, 20, 23, 5, 5, 14, 20, 8, 5, 18, 15, 15, 6, 19, 8, 15, 22, 5, 18, 5, 4, 6, 15, 18, 1, 14, 9, 14, 19, 20, 1, 14, 20, 12, 9, 11, 5, 1, 2, 12, 21, 5, 2, 15, 20, 20, 12, 5, 1, 14, 4, 4, 1, 18, 20, 5, 4, 1, 23, 1, 25, 1, 7, 1, 9, 14, 23, 9, 20, 8, 1, 3, 21, 18, 22, 9, 14, 7, 6, 12, 9, 7, 8, 20, 9, 20, 23, 1, 19, 20, 8, 5, 16, 15, 12, 9, 3, 5, 16, 1, 20, 18, 15, 12, 19, 14, 15, 15, 16, 9, 14, 7, 9, 14, 20, 15, 16, 5, 15, 16, 12, 5, 19, 23, 9, 14, 4, 15, 23, 19, 20, 8, 5, 16, 1, 20, 18, 15, 12, 19, 4, 9, 4, 14, 15, 20, 13, 1, 20, 20, 5, 18, 8, 15, 23, 5, 22, 5, 18, 15, 14, 12, 25, 20, 8, 5, 20, 8, 15, 21, 7, 8, 20, 16, 15, 12, 9, 3, 5, 13, 1, 20, 20, 5, 18, 5, 4]
+
+atbash_15_mapping = reverse!(collect(1:15))
+caesar_2_26_mapping = circshift(collect(1:26), -2)
+affine_3_5_26_mapping = [6, 9, 12, 15, 18, 21, 24, 1, 4, 7, 10, 13, 16, 19, 22, 25, 2, 5, 8, 11, 14, 17, 20, 23, 26, 3]
+affine_3_5_26_mapping_inv = [8, 17, 26, 9, 18, 1, 10, 19, 2, 11, 20, 3, 12, 21, 4, 13, 22, 5, 14, 23, 6, 15, 24, 7, 16, 25]
+
+caesar_2_26_orwell_tokens = [11, 22, 25, 3, 21, 3, 4, 20, 11, 9, 10, 22, 5, 17, 14, 6, 6, 3, 1, 11, 16, 3, 18, 20, 11, 14, 3, 16, 6, 22, 10, 7, 5, 14, 17, 5, 13, 21, 25, 7, 20, 7, 21, 22, 20, 11, 13, 11, 16, 9, 22, 10, 11, 20, 22, 7, 7, 16, 25, 11, 16, 21, 22, 17, 16, 21, 15, 11, 22, 10, 10, 11, 21, 5, 10, 11, 16, 16, 23, 2, 2, 14, 7, 6, 11, 16, 22, 17, 10, 11, 21, 4, 20, 7, 3, 21, 22, 11, 16, 3, 16, 7, 8, 8, 17, 20, 22, 22, 17, 7, 21, 5, 3, 18, 7, 22, 10, 7, 24, 11, 14, 7, 25, 11, 16, 6, 21, 14, 11, 18, 18, 7, 6, 19, 23, 11, 5, 13, 14, 1, 22, 10, 20, 17, 23, 9, 10, 22, 10, 7, 9, 14, 3, 21, 
+21, 6, 17, 17, 20, 21, 17, 8, 24, 11, 5, 22, 17, 20, 1, 15, 3, 16, 21, 11, 17, 16, 21, 22, 10, 17, 23, 9, 10, 16, 
+17, 22, 19, 23, 11, 5, 13, 14, 1, 7, 16, 17, 23, 9, 10, 22, 17, 18, 20, 7, 24, 7, 16, 22, 3, 21, 25, 11, 20, 14, 17, 8, 9, 20, 11, 22, 22, 1, 6, 23, 21, 22, 8, 20, 17, 15, 7, 16, 22, 7, 20, 11, 16, 9, 3, 14, 17, 16, 9, 25, 11, 22, 10, 10, 11, 15, 22, 10, 7, 10, 3, 14, 14, 25, 3, 1, 21, 15, 7, 14, 22, 17, 8, 4, 17, 11, 14, 7, 6, 5, 3, 4, 4, 
+3, 9, 7, 3, 16, 6, 17, 14, 6, 20, 3, 9, 15, 3, 22, 21, 3, 22, 17, 16, 7, 7, 16, 6, 17, 8, 11, 22, 3, 5, 17, 14, 17, 23, 20, 7, 6, 18, 17, 21, 22, 7, 20, 22, 17, 17, 14, 3, 20, 9, 7, 8, 17, 20, 11, 16, 6, 17, 17, 20, 6, 11, 21, 18, 14, 3, 1, 10, 3, 6, 4, 7, 7, 16, 22, 3, 5, 13, 7, 6, 22, 17, 22, 10, 7, 25, 3, 14, 14, 11, 22, 6, 7, 18, 11, 5, 22, 7, 6, 21, 11, 15, 18, 14, 1, 3, 16, 7, 16, 17, 20, 15, 17, 23, 21, 8, 3, 5, 7, 15, 17, 20, 7, 22, 10, 3, 16, 
+3, 15, 7, 22, 20, 7, 25, 11, 6, 7, 22, 10, 7, 8, 3, 5, 7, 17, 8, 3, 15, 3, 16, 17, 8, 3, 4, 17, 23, 22, 8, 17, 20, 22, 1, 8, 11, 24, 7, 25, 11, 22, 10, 3, 10, 7, 3, 24, 1, 4, 14, 3, 5, 13, 15, 17, 23, 21, 22, 3, 5, 10, 7, 3, 16, 6, 20, 23, 9, 9, 7, 6, 14, 1, 10, 3, 16, 6, 21, 17, 15, 7, 8, 7, 3, 22, 23, 20, 7, 21, 25, 11, 16, 21, 22, 17, 16, 15, 3, 6, 7, 8, 17, 20, 22, 10, 7, 21, 22, 3, 11, 20, 21, 11, 22, 25, 3, 21, 16, 17, 23, 21, 7, 22, 20, 1, 11, 16, 9, 22, 10, 7, 14, 11, 8, 22, 7, 24, 7, 16, 3, 22, 22, 10, 7, 4, 7, 21, 22, 17, 8, 22, 11, 15, 7, 21, 11, 22, 25, 3, 21, 21, 7, 14, 6, 17, 15, 25, 17, 20, 13, 11, 16, 9, 3, 16, 6, 3, 22, 18, 20, 7, 21, 7, 16, 22, 22, 10, 7, 7, 14, 7, 5, 22, 20, 11, 5, 5, 23, 20, 20, 7, 16, 22, 25, 3, 21, 5, 23, 22, 17, 8, 8, 6, 23, 20, 11, 16, 9, 6, 3, 1, 14, 11, 9, 10, 22, 10, 17, 23, 20, 21, 11, 22, 25, 3, 21, 18, 3, 20, 22, 17, 8, 22, 10, 7, 7, 5, 17, 16, 17, 15, 
+1, 6, 20, 11, 24, 7, 11, 16, 18, 20, 7, 18, 3, 20, 3, 22, 11, 17, 16, 8, 17, 20, 10, 3, 22, 7, 25, 7, 7, 13, 22, 10, 7, 8, 14, 3, 22, 25, 3, 21, 21, 7, 24, 7, 16, 8, 14, 11, 9, 10, 22, 21, 23, 18, 3, 16, 6, 25, 11, 16, 21, 22, 17, 16, 25, 10, 17, 25, 3, 21, 22, 10, 11, 20, 22, 1, 16, 11, 16, 7, 3, 16, 6, 10, 3, 6, 3, 24, 3, 20, 11, 5, 17, 21, 7, 23, 14, 5, 7, 20, 3, 4, 17, 24, 7, 10, 11, 21, 20, 11, 9, 10, 22, 3, 16, 13, 14, 7, 25, 7, 16, 22, 21, 14, 17, 25, 14, 1, 20, 7, 21, 22, 11, 16, 9, 21, 7, 24, 7, 20, 3, 14, 22, 11, 15, 7, 21, 17, 16, 22, 10, 7, 25, 3, 1, 17, 16, 7, 3, 5, 10, 14, 3, 16, 6, 11, 16, 9, 17, 18, 18, 17, 21, 11, 22, 7, 22, 10, 7, 14, 11, 8, 22, 21, 10, 3, 8, 22, 22, 10, 7, 18, 17, 21, 22, 7, 20, 25, 11, 22, 10, 22, 10, 7, 7, 16, 17, 20, 15, 17, 23, 21, 8, 3, 5, 7, 9, 3, 2, 7, 6, 8, 20, 17, 15, 22, 10, 7, 25, 3, 14, 14, 11, 22, 25, 3, 21, 17, 16, 7, 17, 8, 22, 10, 17, 21, 7, 18, 11, 5, 22, 23, 20, 7, 21, 25, 10, 11, 5, 10, 3, 20, 7, 21, 17, 5, 17, 16, 22, 20, 11, 24, 7, 6, 22, 10, 3, 22, 22, 10, 7, 7, 1, 7, 21, 8, 17, 14, 14, 17, 25, 1, 17, 23, 3, 4, 17, 23, 22, 25, 10, 7, 16, 1, 17, 23, 15, 17, 24, 7, 4, 11, 9, 4, 20, 17, 22, 10, 7, 20, 11, 21, 25, 3, 22, 5, 10, 11, 16, 9, 1, 17, 23, 22, 10, 7, 5, 3, 18, 22, 11, 17, 16, 4, 7, 16, 7, 3, 22, 10, 11, 22, 20, 3, 16, 11, 16, 21, 11, 6, 7, 22, 10, 7, 8, 14, 3, 22, 3, 8, 20, 23, 11, 22, 1, 24, 17, 11, 5, 7, 25, 3, 21, 20, 7, 3, 6, 11, 16, 9, 17, 23, 22, 3, 14, 11, 21, 22, 17, 8, 8, 11, 9, 23, 20, 7, 21, 25, 10, 11, 5, 10, 10, 3, 6, 21, 17, 15, 7, 22, 10, 11, 16, 9, 22, 17, 6, 17, 25, 11, 22, 10, 22, 10, 7, 18, 20, 17, 6, 23, 5, 22, 11, 17, 16, 17, 8, 18, 11, 9, 11, 20, 17, 16, 22, 10, 7, 24, 17, 11, 5, 7, 5, 3, 15, 7, 8, 20, 17, 15, 3, 16, 17, 4, 14, 17, 16, 9, 15, 7, 22, 3, 14, 18, 14, 3, 19, 23, 7, 14, 11, 13, 7, 3, 6, 23, 14, 14, 7, 6, 15, 11, 20, 20, 17, 20, 25, 10, 11, 5, 10, 8, 17, 20, 15, 7, 6, 18, 3, 20, 22, 17, 8, 22, 10, 7, 21, 23, 20, 8, 3, 5, 7, 17, 8, 22, 10, 7, 20, 11, 9, 10, 22, 10, 3, 16, 6, 25, 3, 14, 14, 25, 11, 16, 21, 22, 17, 16, 22, 23, 20, 16, 7, 6, 3, 21, 25, 11, 22, 5, 10, 3, 16, 6, 22, 10, 7, 24, 17, 11, 5, 7, 21, 3, 16, 13, 21, 17, 15, 7, 25, 10, 3, 22, 22, 10, 17, 23, 9, 10, 22, 10, 7, 25, 17, 20, 6, 21, 25, 7, 20, 7, 21, 22, 11, 14, 14, 6, 11, 21, 22, 11, 16, 9, 23, 11, 21, 10, 3, 4, 14, 7, 22, 10, 7, 11, 16, 21, 22, 20, 23, 15, 7, 16, 22, 22, 10, 7, 22, 7, 14, 7, 21, 5, 20, 7, 7, 16, 11, 22, 25, 3, 21, 5, 3, 14, 14, 7, 6, 5, 17, 23, 14, 6, 4, 7, 6, 11, 15, 15, 7, 6, 4, 23, 
+22, 22, 10, 7, 20, 7, 25, 3, 21, 16, 17, 25, 3, 1, 17, 8, 21, 10, 23, 22, 22, 11, 16, 9, 11, 22, 17, 8, 8, 5, 17, 
+15, 18, 14, 7, 22, 7, 14, 1, 10, 7, 15, 17, 24, 7, 6, 17, 24, 7, 20, 22, 17, 22, 10, 7, 25, 11, 16, 6, 17, 25, 3, 
+21, 15, 3, 14, 14, 11, 21, 10, 8, 20, 3, 11, 14, 8, 11, 9, 23, 20, 7, 22, 10, 7, 15, 7, 3, 9, 20, 7, 16, 7, 21, 21, 17, 8, 10, 11, 21, 4, 17, 6, 1, 15, 7, 20, 7, 14, 1, 7, 15, 18, 10, 3, 21, 11, 2, 7, 6, 4, 1, 22, 10, 7, 4, 14, 
+23, 7, 17, 24, 7, 20, 3, 14, 14, 21, 25, 10, 11, 5, 10, 25, 7, 20, 7, 22, 10, 7, 23, 16, 11, 8, 17, 20, 15, 17, 8, 22, 10, 7, 18, 3, 20, 22, 1, 10, 11, 21, 10, 3, 11, 20, 25, 3, 21, 24, 7, 20, 1, 8, 3, 11, 20, 10, 11, 21, 8, 3, 
+5, 7, 16, 3, 22, 23, 20, 3, 14, 14, 1, 21, 3, 16, 9, 23, 11, 16, 7, 10, 11, 21, 21, 13, 11, 16, 20, 17, 23, 9, 10, 7, 16, 7, 6, 4, 1, 5, 17, 3, 20, 21, 7, 21, 17, 3, 18, 3, 16, 6, 4, 14, 23, 16, 22, 20, 3, 2, 17, 20, 4, 14, 3, 6, 7, 21, 3, 16, 6, 22, 10, 7, 5, 17, 14, 6, 17, 8, 22, 10, 7, 25, 11, 16, 22, 7, 20, 22, 10, 3, 22, 10, 3, 6, 12, 
+23, 21, 22, 7, 16, 6, 7, 6, 17, 23, 22, 21, 11, 6, 7, 7, 24, 7, 16, 22, 10, 20, 17, 23, 9, 10, 22, 10, 7, 21, 10, 
+23, 22, 25, 11, 16, 6, 17, 25, 18, 3, 16, 7, 22, 10, 7, 25, 17, 20, 14, 6, 14, 17, 17, 13, 7, 6, 5, 17, 14, 6, 6, 
+17, 25, 16, 11, 16, 22, 10, 7, 21, 22, 20, 7, 7, 22, 14, 11, 22, 22, 14, 7, 7, 6, 6, 11, 7, 21, 17, 8, 25, 11, 16, 6, 25, 7, 20, 7, 25, 10, 11, 20, 14, 11, 16, 9, 6, 23, 21, 22, 3, 16, 6, 22, 17, 20, 16, 18, 3, 18, 7, 20, 11, 16, 22, 17, 21, 18, 11, 20, 3, 14, 21, 3, 16, 6, 22, 10, 17, 23, 9, 10, 22, 10, 7, 21, 23, 16, 25, 3, 21, 21, 10, 11, 16, 11, 16, 9, 3, 16, 6, 22, 10, 7, 21, 13, 1, 3, 10, 3, 20, 21, 10, 4, 14, 23, 7, 22, 10, 7, 20, 7, 21, 7, 7, 15, 7, 6, 22, 17, 4, 7, 16, 17, 5, 17, 14, 17, 23, 20, 11, 16, 3, 16, 1, 22, 10, 11, 16, 9, 7, 26, 5, 7, 18, 22, 22, 10, 7, 18, 17, 21, 22, 7, 20, 21, 22, 10, 3, 22, 25, 7, 20, 7, 18, 14, 3, 21, 22, 7, 20, 7, 6, 7, 24, 7, 20, 1, 
+25, 10, 7, 20, 7, 22, 10, 7, 4, 14, 3, 5, 13, 15, 17, 23, 21, 22, 3, 5, 10, 11, 17, 6, 8, 3, 5, 7, 9, 3, 2, 7, 6, 
+6, 17, 25, 16, 8, 20, 17, 15, 7, 24, 7, 20, 1, 5, 17, 15, 15, 3, 16, 6, 11, 16, 9, 5, 17, 20, 16, 7, 20, 22, 10, 7, 20, 7, 25, 3, 21, 17, 16, 7, 17, 16, 22, 10, 7, 10, 17, 23, 21, 7, 8, 20, 17, 16, 22, 11, 15, 15, 7, 6, 11, 3, 22, 7, 14, 1, 17, 18, 18, 17, 21, 11, 22, 7, 4, 11, 9, 4, 20, 17, 22, 10, 7, 20, 11, 21, 25, 3, 22, 5, 10, 11, 16, 
+9, 1, 17, 23, 22, 10, 7, 5, 3, 18, 22, 11, 17, 16, 21, 3, 11, 6, 25, 10, 11, 14, 7, 22, 10, 7, 6, 3, 20, 13, 7, 1, 7, 21, 14, 17, 17, 13, 7, 6, 6, 7, 7, 18, 11, 16, 22, 17, 25, 11, 16, 21, 22, 17, 16, 21, 17, 25, 16, 6, 17, 25, 
+16, 3, 22, 21, 22, 20, 7, 7, 22, 14, 7, 24, 7, 14, 3, 16, 17, 22, 10, 7, 20, 18, 17, 21, 22, 7, 20, 22, 17, 20, 16, 3, 22, 17, 16, 7, 5, 17, 20, 16, 7, 20, 8, 14, 3, 18, 18, 7, 6, 8, 11, 22, 8, 23, 14, 14, 1, 11, 16, 22, 10, 7, 
+25, 11, 16, 6, 3, 14, 22, 7, 20, 16, 3, 22, 7, 14, 1, 5, 17, 24, 7, 20, 11, 16, 9, 3, 16, 6, 23, 16, 5, 17, 24, 7, 20, 11, 16, 9, 22, 10, 7, 21, 11, 16, 9, 14, 7, 25, 17, 20, 6, 11, 16, 9, 21, 17, 5, 11, 16, 22, 10, 7, 8, 3, 20, 6, 11, 21, 22, 3, 16, 5, 7, 3, 10, 7, 14, 11, 5, 17, 18, 22, 7, 20, 21, 13, 11, 15, 15, 7, 6, 6, 17, 25, 16, 4, 7, 22, 25, 7, 7, 16, 22, 10, 7, 20, 17, 17, 8, 21, 10, 17, 24, 7, 20, 7, 6, 8, 17, 20, 3, 16, 11, 16, 21, 22, 3, 16, 22, 14, 11, 13, 7, 3, 4, 14, 23, 7, 4, 17, 22, 22, 14, 7, 3, 16, 6, 6, 3, 20, 22, 7, 6, 3, 25, 3, 1, 3, 9, 3, 11, 16, 25, 11, 22, 10, 3, 5, 23, 20, 24, 11, 16, 9, 8, 14, 11, 9, 10, 22, 11, 22, 25, 3, 21, 22, 10, 7, 18, 17, 14, 11, 5, 7, 18, 3, 22, 20, 17, 14, 21, 16, 17, 17, 18, 11, 16, 9, 11, 16, 22, 17, 18, 7, 17, 18, 14, 7, 21, 25, 11, 16, 6, 17, 25, 21, 22, 10, 7, 18, 3, 22, 20, 17, 14, 21, 6, 11, 6, 16, 17, 22, 15, 3, 22, 22, 7, 20, 10, 17, 25, 
+7, 24, 7, 20, 17, 16, 14, 1, 22, 10, 7, 22, 10, 17, 23, 9, 10, 22, 18, 17, 14, 11, 5, 7, 15, 3, 22, 22, 7, 20, 7, 
+6]
+affine_3_5_26_orwell_tokens = [4, 11, 20, 6, 8, 6, 9, 5, 4, 24, 1, 11, 12, 22, 13, 15, 15, 6, 26, 4, 19, 6, 25, 5, 4, 13, 6, 19, 15, 11, 1, 18, 
+12, 13, 22, 12, 10, 8, 20, 18, 5, 18, 8, 11, 5, 4, 10, 4, 19, 24, 11, 1, 4, 5, 11, 18, 18, 19, 20, 4, 19, 8, 11, 22, 19, 8, 16, 4, 11, 1, 1, 4, 8, 12, 1, 4, 19, 19, 14, 3, 3, 13, 18, 15, 4, 19, 11, 22, 1, 4, 8, 9, 5, 18, 6, 8, 11, 4, 19, 6, 19, 18, 21, 21, 22, 5, 11, 11, 22, 18, 8, 12, 6, 25, 18, 11, 1, 18, 17, 4, 13, 18, 20, 4, 19, 15, 8, 
+13, 4, 25, 25, 18, 15, 2, 14, 4, 12, 10, 13, 26, 11, 1, 5, 22, 14, 24, 1, 11, 1, 18, 24, 13, 6, 8, 8, 15, 22, 22, 
+5, 8, 22, 21, 17, 4, 12, 11, 22, 5, 26, 16, 6, 19, 8, 4, 22, 19, 8, 11, 1, 22, 14, 24, 1, 19, 22, 11, 2, 14, 4, 12, 10, 13, 26, 18, 19, 22, 14, 24, 1, 11, 22, 25, 5, 18, 17, 18, 19, 11, 6, 8, 20, 4, 5, 13, 22, 21, 24, 5, 4, 11, 
+11, 26, 15, 14, 8, 11, 21, 5, 22, 16, 18, 19, 11, 18, 5, 4, 19, 24, 6, 13, 22, 19, 24, 20, 4, 11, 1, 1, 4, 16, 11, 1, 18, 1, 6, 13, 13, 20, 6, 26, 8, 16, 18, 13, 11, 22, 21, 9, 22, 4, 13, 18, 15, 12, 6, 9, 9, 6, 24, 18, 6, 19, 15, 22, 13, 15, 5, 6, 24, 16, 6, 11, 8, 6, 11, 22, 19, 18, 18, 19, 15, 22, 21, 4, 11, 6, 12, 22, 13, 22, 14, 5, 18, 15, 25, 22, 8, 11, 18, 5, 11, 22, 22, 13, 6, 5, 24, 18, 21, 22, 5, 4, 19, 15, 22, 22, 5, 15, 4, 8, 25, 13, 6, 26, 1, 6, 15, 9, 18, 18, 19, 11, 6, 12, 10, 18, 15, 11, 22, 11, 1, 18, 20, 6, 13, 13, 4, 11, 15, 18, 25, 4, 12, 11, 18, 15, 8, 4, 16, 25, 13, 26, 6, 19, 18, 19, 22, 5, 16, 22, 14, 8, 21, 6, 12, 18, 16, 22, 5, 18, 11, 1, 6, 19, 6, 16, 18, 11, 5, 18, 20, 4, 15, 18, 11, 1, 18, 21, 6, 12, 18, 22, 21, 6, 16, 6, 19, 22, 21, 6, 9, 22, 14, 11, 21, 22, 5, 11, 26, 21, 4, 17, 18, 20, 4, 11, 1, 6, 1, 18, 6, 17, 26, 9, 13, 6, 12, 10, 16, 22, 14, 8, 11, 6, 12, 1, 18, 6, 19, 15, 5, 14, 24, 24, 18, 15, 13, 26, 1, 6, 19, 15, 8, 22, 16, 18, 21, 18, 6, 11, 14, 5, 18, 8, 20, 4, 19, 8, 11, 22, 19, 16, 6, 15, 18, 21, 22, 5, 11, 1, 18, 8, 11, 6, 4, 5, 8, 4, 11, 20, 6, 8, 19, 22, 14, 8, 18, 11, 5, 26, 
+4, 19, 24, 11, 1, 18, 13, 4, 21, 11, 18, 17, 18, 19, 6, 11, 11, 1, 18, 9, 18, 8, 11, 22, 21, 11, 4, 16, 18, 8, 4, 
+11, 20, 6, 8, 8, 18, 13, 15, 22, 16, 20, 22, 5, 10, 4, 19, 24, 6, 19, 15, 6, 11, 25, 5, 18, 8, 18, 19, 11, 11, 1, 
+18, 18, 13, 18, 12, 11, 5, 4, 12, 12, 14, 5, 5, 18, 19, 11, 20, 6, 8, 12, 14, 11, 22, 21, 21, 15, 14, 5, 4, 19, 24, 15, 6, 26, 13, 4, 24, 1, 11, 1, 22, 14, 5, 8, 4, 11, 20, 6, 8, 25, 6, 5, 11, 22, 21, 11, 1, 18, 18, 12, 22, 19, 
+22, 16, 26, 15, 5, 4, 17, 18, 4, 19, 25, 5, 18, 25, 6, 5, 6, 11, 4, 22, 19, 21, 22, 5, 1, 6, 11, 18, 20, 18, 18, 10, 11, 1, 18, 21, 13, 6, 11, 20, 6, 8, 8, 18, 17, 18, 19, 21, 13, 4, 24, 1, 11, 8, 14, 25, 6, 19, 15, 20, 4, 19, 8, 11, 22, 19, 20, 1, 22, 20, 6, 8, 11, 1, 4, 5, 11, 26, 19, 4, 19, 18, 6, 19, 15, 1, 6, 15, 6, 17, 6, 5, 4, 12, 22, 8, 18, 14, 13, 12, 18, 5, 6, 9, 22, 17, 18, 1, 4, 8, 5, 4, 24, 1, 11, 6, 19, 10, 13, 18, 20, 18, 19, 11, 8, 13, 
+22, 20, 13, 26, 5, 18, 8, 11, 4, 19, 24, 8, 18, 17, 18, 5, 6, 13, 11, 4, 16, 18, 8, 22, 19, 11, 1, 18, 20, 6, 26, 
+22, 19, 18, 6, 12, 1, 13, 6, 19, 15, 4, 19, 24, 22, 25, 25, 22, 8, 4, 11, 18, 11, 1, 18, 13, 4, 21, 11, 8, 1, 6, 21, 11, 11, 1, 18, 25, 22, 8, 11, 18, 5, 20, 4, 11, 1, 11, 1, 18, 18, 19, 22, 5, 16, 22, 14, 8, 21, 6, 12, 18, 24, 
+6, 3, 18, 15, 21, 5, 22, 16, 11, 1, 18, 20, 6, 13, 13, 4, 11, 20, 6, 8, 22, 19, 18, 22, 21, 11, 1, 22, 8, 18, 25, 
+4, 12, 11, 14, 5, 18, 8, 20, 1, 4, 12, 1, 6, 5, 18, 8, 22, 12, 22, 19, 11, 5, 4, 17, 18, 15, 11, 1, 6, 11, 11, 1, 
+18, 18, 26, 18, 8, 21, 22, 13, 13, 22, 20, 26, 22, 14, 6, 9, 22, 14, 11, 20, 1, 18, 19, 26, 22, 14, 16, 22, 17, 18, 9, 4, 24, 9, 5, 22, 11, 1, 18, 5, 4, 8, 20, 6, 11, 12, 1, 4, 19, 24, 26, 22, 14, 11, 1, 18, 12, 6, 25, 11, 4, 22, 19, 9, 18, 19, 18, 6, 11, 1, 4, 11, 5, 6, 19, 4, 19, 8, 4, 15, 18, 11, 1, 18, 21, 13, 6, 11, 6, 21, 5, 14, 4, 11, 26, 17, 22, 4, 12, 18, 20, 6, 8, 5, 18, 6, 15, 4, 19, 24, 22, 14, 11, 6, 13, 4, 8, 11, 22, 21, 21, 4, 24, 14, 5, 18, 8, 20, 1, 4, 12, 1, 1, 6, 15, 8, 22, 16, 18, 11, 1, 4, 19, 24, 11, 22, 15, 22, 20, 4, 11, 1, 11, 1, 18, 25, 5, 22, 15, 14, 12, 11, 4, 22, 19, 22, 21, 25, 4, 24, 4, 5, 22, 19, 11, 1, 18, 17, 22, 4, 12, 18, 12, 6, 16, 18, 21, 5, 22, 16, 6, 19, 22, 9, 13, 22, 19, 24, 16, 18, 11, 6, 13, 25, 13, 6, 2, 14, 18, 13, 4, 10, 18, 6, 15, 14, 13, 13, 18, 15, 16, 4, 5, 5, 22, 5, 20, 1, 4, 12, 1, 21, 22, 5, 16, 18, 15, 25, 6, 5, 11, 22, 21, 11, 1, 18, 8, 14, 5, 
+21, 6, 12, 18, 22, 21, 11, 1, 18, 5, 4, 24, 1, 11, 1, 6, 19, 15, 20, 6, 13, 13, 20, 4, 19, 8, 11, 22, 19, 11, 14, 
+5, 19, 18, 15, 6, 8, 20, 4, 11, 12, 1, 6, 19, 15, 11, 1, 18, 17, 22, 4, 12, 18, 8, 6, 19, 10, 8, 22, 16, 18, 20, 1, 6, 11, 11, 1, 22, 14, 24, 1, 11, 1, 18, 20, 22, 5, 15, 8, 20, 18, 5, 18, 8, 11, 4, 13, 13, 15, 4, 8, 11, 4, 19, 
+24, 14, 4, 8, 1, 6, 9, 13, 18, 11, 1, 18, 4, 19, 8, 11, 5, 14, 16, 18, 19, 11, 11, 1, 18, 11, 18, 13, 18, 8, 12, 5, 18, 18, 19, 4, 11, 20, 6, 8, 12, 6, 13, 13, 18, 15, 12, 22, 14, 13, 15, 9, 18, 15, 4, 16, 16, 18, 15, 9, 14, 11, 11, 1, 18, 5, 18, 20, 6, 8, 19, 22, 20, 6, 26, 22, 21, 8, 1, 14, 11, 11, 4, 19, 24, 4, 11, 22, 21, 21, 12, 22, 16, 25, 13, 18, 11, 18, 13, 26, 1, 18, 16, 22, 17, 18, 15, 22, 17, 18, 5, 11, 22, 11, 1, 18, 20, 4, 19, 15, 22, 20, 
+6, 8, 16, 6, 13, 13, 4, 8, 1, 21, 5, 6, 4, 13, 21, 4, 24, 14, 5, 18, 11, 1, 18, 16, 18, 6, 24, 5, 18, 19, 18, 8, 8, 22, 21, 1, 4, 8, 9, 22, 15, 26, 16, 18, 5, 18, 13, 26, 18, 16, 25, 1, 6, 8, 4, 3, 18, 15, 9, 26, 11, 1, 18, 9, 13, 14, 18, 22, 17, 18, 5, 6, 13, 13, 8, 20, 1, 4, 12, 1, 20, 18, 5, 18, 11, 1, 18, 14, 19, 4, 21, 22, 5, 16, 22, 21, 11, 1, 18, 25, 6, 5, 11, 26, 1, 4, 8, 1, 6, 4, 5, 20, 6, 8, 17, 18, 5, 26, 21, 6, 4, 5, 1, 4, 8, 21, 6, 12, 18, 19, 6, 11, 14, 5, 6, 13, 13, 26, 8, 6, 19, 24, 14, 4, 19, 18, 1, 4, 8, 8, 10, 4, 19, 5, 22, 14, 24, 1, 18, 19, 18, 15, 9, 26, 12, 22, 6, 5, 8, 18, 8, 22, 6, 25, 6, 19, 15, 9, 13, 14, 19, 11, 5, 6, 3, 22, 5, 9, 13, 6, 15, 18, 8, 6, 19, 15, 11, 1, 18, 12, 22, 13, 15, 22, 21, 11, 1, 18, 20, 4, 19, 11, 18, 5, 11, 1, 6, 11, 1, 6, 15, 7, 14, 8, 
+11, 18, 19, 15, 18, 15, 22, 14, 11, 8, 4, 15, 18, 18, 17, 18, 19, 11, 1, 5, 22, 14, 24, 1, 11, 1, 18, 8, 1, 14, 11, 20, 4, 19, 15, 22, 20, 25, 6, 19, 18, 11, 1, 18, 20, 22, 5, 13, 15, 13, 22, 22, 10, 18, 15, 12, 22, 13, 15, 15, 
+22, 20, 19, 4, 19, 11, 1, 18, 8, 11, 5, 18, 18, 11, 13, 4, 11, 11, 13, 18, 18, 15, 15, 4, 18, 8, 22, 21, 20, 4, 19, 15, 20, 18, 5, 18, 20, 1, 4, 5, 13, 4, 19, 24, 15, 14, 8, 11, 6, 19, 15, 11, 22, 5, 19, 25, 6, 25, 18, 5, 4, 19, 11, 22, 8, 25, 4, 5, 6, 13, 8, 6, 19, 15, 11, 1, 22, 14, 24, 1, 11, 1, 18, 8, 14, 19, 20, 6, 8, 8, 1, 4, 19, 4, 19, 24, 6, 19, 15, 11, 1, 18, 8, 10, 26, 6, 1, 6, 5, 8, 1, 9, 13, 14, 18, 11, 1, 18, 5, 18, 8, 18, 18, 16, 18, 15, 
+11, 22, 9, 18, 19, 22, 12, 22, 13, 22, 14, 5, 4, 19, 6, 19, 26, 11, 1, 4, 19, 24, 18, 23, 12, 18, 25, 11, 11, 1, 18, 25, 22, 8, 11, 18, 5, 8, 11, 1, 6, 11, 20, 18, 5, 18, 25, 13, 6, 8, 11, 18, 5, 18, 15, 18, 17, 18, 5, 26, 20, 1, 18, 5, 18, 11, 1, 18, 9, 13, 6, 12, 10, 16, 22, 14, 8, 11, 6, 12, 1, 4, 22, 15, 21, 6, 12, 18, 24, 6, 3, 18, 15, 15, 22, 20, 19, 21, 5, 22, 16, 18, 17, 18, 5, 26, 12, 22, 16, 16, 6, 19, 15, 4, 19, 24, 12, 22, 5, 19, 18, 5, 11, 1, 18, 5, 18, 20, 6, 8, 22, 19, 18, 22, 19, 11, 1, 18, 1, 22, 14, 8, 18, 21, 5, 22, 19, 11, 4, 16, 16, 18, 15, 4, 6, 11, 18, 13, 26, 22, 25, 25, 22, 8, 4, 11, 18, 9, 4, 24, 9, 5, 22, 11, 1, 18, 5, 4, 8, 20, 6, 11, 12, 1, 4, 19, 24, 26, 22, 14, 11, 1, 18, 12, 6, 25, 11, 4, 22, 19, 8, 6, 4, 15, 20, 1, 4, 13, 18, 11, 1, 18, 15, 6, 5, 10, 18, 
+26, 18, 8, 13, 22, 22, 10, 18, 15, 15, 18, 18, 25, 4, 19, 11, 22, 20, 4, 19, 8, 11, 22, 19, 8, 22, 20, 19, 15, 22, 20, 19, 6, 11, 8, 11, 5, 18, 18, 11, 13, 18, 17, 18, 13, 6, 19, 22, 11, 1, 18, 5, 25, 22, 8, 11, 18, 5, 11, 22, 5, 19, 6, 11, 22, 19, 18, 12, 22, 5, 19, 18, 5, 21, 13, 6, 25, 25, 18, 15, 21, 4, 11, 21, 14, 13, 13, 26, 4, 19, 11, 1, 18, 20, 4, 19, 15, 6, 13, 11, 18, 5, 19, 6, 11, 18, 13, 26, 12, 22, 17, 18, 5, 4, 19, 24, 6, 19, 15, 14, 19, 
+12, 22, 17, 18, 5, 4, 19, 24, 11, 1, 18, 8, 4, 19, 24, 13, 18, 20, 22, 5, 15, 4, 19, 24, 8, 22, 12, 4, 19, 11, 1, 
+18, 21, 6, 5, 15, 4, 8, 11, 6, 19, 12, 18, 6, 1, 18, 13, 4, 12, 22, 25, 11, 18, 5, 8, 10, 4, 16, 16, 18, 15, 15, 22, 20, 19, 9, 18, 11, 20, 18, 18, 19, 11, 1, 18, 5, 22, 22, 21, 8, 1, 22, 17, 18, 5, 18, 15, 21, 22, 5, 6, 19, 4, 
+19, 8, 11, 6, 19, 11, 13, 4, 10, 18, 6, 9, 13, 14, 18, 9, 22, 11, 11, 13, 18, 6, 19, 15, 15, 6, 5, 11, 18, 15, 6, 
+20, 6, 26, 6, 24, 6, 4, 19, 20, 4, 11, 1, 6, 12, 14, 5, 17, 4, 19, 24, 21, 13, 4, 24, 1, 11, 4, 11, 20, 6, 8, 11, 
+1, 18, 25, 22, 13, 4, 12, 18, 25, 6, 11, 5, 22, 13, 8, 19, 22, 22, 25, 4, 19, 24, 4, 19, 11, 22, 25, 18, 22, 25, 13, 18, 8, 20, 4, 19, 15, 22, 20, 8, 11, 1, 18, 25, 6, 11, 5, 22, 13, 8, 15, 4, 15, 19, 22, 11, 16, 6, 11, 11, 18, 
+5, 1, 22, 20, 18, 17, 18, 5, 22, 19, 13, 26, 11, 1, 18, 11, 1, 22, 14, 24, 1, 11, 25, 22, 13, 4, 12, 18, 16, 6, 11, 11, 18, 5, 18, 15]
+
+freq_match_orwell_mapping = [15, 2, 23, 12, 5, 7, 16, 18, 9, 26, 11, 4, 21, 14, 1, 25, 24, 19, 8, 20, 3, 22, 6, 17, 13, 10]
+
+
+
+
+
+
+vigenere_ABBA_mappings = [[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26], [2, 3, 4, 5, 6, 
+7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 1], [2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 
+12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 1], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26]]
+
+periodic_affine_123_765_mappings = [[8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 1, 2, 3, 4, 5, 6, 7], [7, 10, 13, 16, 
+19, 22, 25, 2, 5, 8, 11, 14, 17, 20, 23, 26, 3, 6, 9, 12, 15, 18, 21, 24, 1, 4], [6, 11, 16, 21, 26, 5, 10, 15, 20, 25, 4, 9, 14, 19, 24, 3, 8, 13, 18, 23, 2, 7, 12, 17, 22, 1]]
+
+vigenere_ABBA_orwell_tokens = [9, 21, 24, 1, 19, 2, 3, 18, 9, 8, 9, 20, 3, 16, 13, 4, 4, 2, 26, 9, 14, 2, 17, 18, 9, 13, 2, 14, 4, 21, 9, 5, 3, 
+13, 16, 3, 11, 20, 24, 5, 18, 6, 20, 20, 18, 10, 12, 9, 14, 8, 21, 8, 9, 19, 21, 5, 5, 15, 24, 9, 14, 20, 21, 15, 
+14, 20, 14, 9, 20, 9, 9, 9, 19, 4, 9, 9, 14, 15, 22, 26, 26, 13, 6, 4, 9, 15, 21, 15, 8, 10, 20, 2, 18, 6, 2, 19, 
+20, 10, 15, 1, 14, 6, 7, 6, 15, 19, 21, 20, 15, 6, 20, 3, 1, 17, 6, 20, 8, 6, 23, 9, 12, 6, 24, 9, 14, 5, 20, 12, 
+9, 17, 17, 5, 4, 18, 22, 9, 3, 12, 13, 25, 20, 9, 19, 15, 21, 8, 9, 20, 8, 6, 8, 12, 1, 20, 20, 4, 15, 16, 19, 19, 15, 7, 23, 9, 3, 21, 16, 18, 25, 14, 2, 14, 19, 10, 16, 14, 19, 21, 9, 15, 21, 8, 9, 14, 15, 21, 18, 21, 9, 4, 12, 12, 25, 6, 15, 15, 21, 8, 9, 20, 15, 17, 19, 5, 22, 6, 15, 20, 1, 20, 24, 9, 18, 13, 16, 6, 7, 19, 10, 20, 20, 26, 5, 21, 19, 21, 7, 18, 15, 14, 6, 14, 20, 6, 19, 9, 14, 8, 2, 12, 15, 15, 8, 23, 9, 21, 9, 8, 9, 14, 21, 8, 5, 9, 2, 12, 12, 24, 2, 25, 19, 14, 6, 12, 20, 16, 7, 2, 15, 10, 13, 5, 4, 4, 2, 2, 2, 2, 8, 5, 1, 15, 5, 15, 12, 5, 19, 1, 7, 14, 2, 20, 19, 2, 21, 15, 14, 6, 6, 14, 4, 16, 7, 9, 20, 2, 4, 15, 12, 16, 22, 18, 5, 5, 17, 15, 19, 21, 
+6, 18, 20, 16, 16, 12, 1, 19, 8, 5, 6, 16, 19, 9, 14, 5, 16, 15, 18, 5, 10, 19, 16, 13, 2, 25, 8, 2, 5, 2, 5, 6, 15, 20, 1, 4, 12, 5, 4, 21, 16, 20, 8, 6, 24, 1, 12, 13, 10, 20, 4, 6, 17, 9, 3, 21, 6, 4, 19, 10, 14, 16, 12, 26, 
+2, 14, 5, 15, 16, 18, 13, 16, 22, 19, 6, 2, 4, 5, 13, 16, 19, 5, 20, 9, 2, 14, 1, 14, 6, 20, 18, 6, 24, 9, 4, 6, 21, 8, 5, 7, 2, 3, 5, 16, 7, 1, 13, 2, 15, 15, 6, 2, 3, 15, 21, 21, 7, 15, 18, 21, 26, 6, 9, 23, 6, 23, 9, 21, 9, 1, 8, 6, 2, 22, 25, 3, 13, 1, 3, 12, 14, 15, 21, 20, 21, 1, 3, 9, 6, 1, 14, 5, 19, 21, 7, 8, 6, 4, 12, 26, 9, 1, 14, 5, 20, 15, 13, 6, 7, 5, 1, 21, 22, 18, 5, 20, 24, 9, 14, 20, 21, 15, 14, 14, 2, 4, 5, 7, 16, 18, 20, 9, 6, 19, 20, 2, 10, 18, 19, 10, 21, 23, 1, 20, 15, 15, 21, 20, 6, 20, 18, 26, 10, 14, 7, 21, 9, 5, 12, 10, 7, 20, 5, 23, 6, 
+14, 1, 21, 21, 8, 5, 3, 6, 19, 20, 16, 7, 20, 9, 14, 6, 19, 9, 21, 24, 1, 19, 20, 6, 12, 4, 16, 14, 23, 15, 19, 12, 9, 14, 8, 2, 14, 4, 2, 21, 16, 18, 6, 20, 5, 14, 21, 21, 8, 5, 6, 13, 5, 3, 21, 19, 9, 3, 4, 22, 18, 18, 6, 15, 
+20, 23, 2, 20, 3, 21, 21, 16, 6, 6, 5, 22, 18, 9, 15, 8, 4, 1, 26, 13, 9, 7, 9, 21, 8, 15, 22, 19, 19, 9, 21, 24, 
+1, 19, 17, 2, 18, 20, 16, 7, 20, 8, 6, 6, 3, 15, 15, 16, 13, 25, 5, 19, 9, 22, 6, 10, 14, 16, 19, 6, 16, 1, 19, 2, 20, 9, 16, 15, 6, 15, 19, 9, 1, 20, 6, 24, 5, 5, 12, 21, 8, 5, 7, 13, 1, 20, 24, 2, 19, 19, 6, 23, 5, 14, 7, 13, 
+9, 7, 9, 21, 19, 21, 17, 2, 14, 4, 24, 10, 14, 19, 21, 16, 14, 23, 9, 16, 23, 1, 20, 21, 8, 9, 19, 21, 25, 14, 10, 15, 5, 1, 15, 5, 8, 1, 5, 2, 22, 1, 19, 10, 3, 15, 20, 6, 21, 12, 4, 6, 18, 1, 3, 16, 22, 5, 9, 10, 19, 18, 10, 8, 8, 20, 2, 15, 11, 12, 6, 24, 5, 14, 21, 20, 12, 15, 24, 13, 25, 18, 6, 20, 20, 9, 15, 8, 19, 5, 23, 6, 18, 1, 13, 21, 9, 13, 6, 20, 15, 14, 21, 9, 5, 23, 2, 26, 15, 14, 6, 2, 3, 8, 13, 2, 14, 4, 10, 15, 7, 15, 17, 17, 15, 19, 
+10, 21, 5, 20, 9, 6, 12, 9, 7, 21, 19, 8, 2, 7, 20, 20, 9, 6, 16, 15, 20, 21, 5, 18, 24, 10, 20, 8, 21, 9, 5, 5, 15, 16, 18, 13, 16, 22, 19, 6, 2, 4, 5, 7, 2, 1, 5, 4, 7, 19, 15, 13, 21, 9, 5, 23, 2, 13, 12, 9, 21, 24, 1, 19, 16, 15, 5, 15, 7, 21, 8, 15, 20, 6, 16, 9, 4, 21, 21, 18, 6, 20, 23, 8, 10, 4, 8, 1, 19, 6, 19, 15, 4, 16, 14, 20, 19, 10, 22, 5, 5, 21, 8, 1, 21, 21, 8, 5, 6, 26, 5, 19, 7, 16, 12, 12, 16, 24, 25, 15, 22, 2, 2, 15, 22, 21, 23, 8, 6, 15, 25, 15, 22, 14, 15, 22, 6, 3, 9, 7, 3, 19, 15, 20, 9, 6, 18, 9, 20, 24, 1, 20, 4, 9, 9, 14, 8, 26, 15, 21, 21, 9, 5, 3, 2, 17, 20, 9, 16, 15, 2, 5, 15, 6, 1, 20, 9, 10, 20, 18, 2, 15, 9, 14, 20, 10, 4, 5, 21, 9, 5, 6, 13, 2, 20, 1, 7, 19, 21, 9, 21, 26, 22, 15, 10, 4, 5, 23, 2, 20, 18, 5, 2, 5, 9, 14, 8, 16, 21, 20, 2, 13, 9, 19, 21, 16, 6, 6, 10, 8, 21, 18, 6, 20, 23, 8, 10, 4, 8, 8, 2, 5, 19, 15, 14, 6, 20, 8, 10, 15, 7, 20, 16, 5, 15, 23, 10, 21, 8, 20, 9, 6, 16, 18, 16, 5, 21, 3, 21, 10, 15, 14, 16, 7, 16, 9, 8, 10, 18, 15, 15, 21, 8, 5, 23, 16, 9, 3, 
+6, 4, 1, 13, 6, 7, 18, 15, 14, 2, 14, 15, 3, 13, 15, 14, 8, 14, 5, 20, 2, 13, 16, 12, 2, 18, 21, 5, 13, 10, 11, 5, 2, 5, 21, 12, 13, 6, 4, 13, 10, 19, 18, 15, 19, 24, 8, 9, 4, 9, 6, 15, 19, 14, 5, 4, 17, 2, 18, 20, 16, 7, 20, 8, 6, 20, 21, 18, 7, 2, 3, 5, 16, 7, 20, 8, 6, 19, 9, 7, 9, 21, 8, 1, 15, 5, 23, 1, 13, 13, 23, 9, 15, 20, 20, 15, 15, 21, 21, 18, 15, 6, 4, 1, 20, 24, 9, 20, 4, 9, 1, 14, 5, 21, 8, 5, 23, 16, 9, 3, 6, 20, 1, 14, 12, 20, 15, 13, 6, 24, 8, 1, 21, 21, 8, 15, 22, 8, 8, 20, 9, 6, 23, 15, 19, 5, 19, 23, 6, 19, 5, 19, 21, 10, 12, 12, 5, 10, 19, 20, 10, 15, 7, 21, 10, 20, 8, 1, 3, 13, 5, 20, 9, 6, 9, 14, 20, 21, 18, 21, 14, 6, 14, 20, 21, 9, 5, 20, 6, 13, 5, 19, 4, 19, 5, 5, 15, 10, 20, 23, 2, 20, 3, 1, 13, 13, 5, 4, 4, 16, 21, 12, 5, 3, 5, 4, 10, 14, 13, 5, 5, 3, 21, 20, 
+21, 9, 5, 18, 6, 24, 1, 19, 15, 16, 23, 1, 26, 16, 6, 19, 9, 22, 20, 20, 10, 15, 7, 9, 21, 16, 6, 6, 4, 16, 13, 16, 13, 6, 20, 5, 13, 26, 8, 5, 14, 16, 22, 5, 5, 16, 22, 5, 19, 21, 15, 20, 9, 6, 23, 9, 15, 5, 15, 23, 2, 20, 13, 
+1, 13, 13, 9, 19, 9, 7, 18, 1, 10, 13, 6, 9, 8, 22, 18, 5, 21, 9, 5, 13, 6, 2, 7, 18, 6, 15, 5, 19, 20, 16, 6, 8, 
+10, 20, 2, 15, 5, 26, 13, 5, 19, 6, 12, 25, 6, 14, 16, 8, 2, 20, 9, 26, 6, 5, 2, 25, 21, 9, 5, 2, 13, 22, 5, 15, 23, 6, 18, 1, 13, 13, 19, 23, 9, 10, 3, 8, 24, 6, 18, 5, 21, 9, 5, 21, 15, 10, 6, 15, 19, 14, 15, 6, 21, 9, 5, 16, 
+2, 19, 20, 25, 9, 10, 19, 8, 2, 10, 18, 23, 2, 20, 22, 5, 19, 26, 6, 1, 10, 19, 8, 9, 20, 7, 1, 3, 6, 15, 1, 20, 22, 19, 1, 12, 13, 26, 19, 1, 15, 8, 21, 9, 15, 6, 8, 9, 20, 20, 11, 9, 15, 19, 15, 21, 8, 9, 5, 14, 6, 5, 2, 25, 4, 16, 1, 18, 20, 6, 19, 15, 2, 17, 1, 14, 5, 3, 12, 21, 15, 21, 18, 1, 1, 16, 18, 2, 13, 2, 4, 5, 20, 2, 14, 4, 21, 9, 5, 3, 16, 13, 4, 15, 7, 21, 8, 5, 24, 10, 14, 20, 6, 19, 20, 8, 2, 21, 8, 1, 5, 11, 21, 19, 21, 6, 14, 4, 6, 
+5, 15, 21, 21, 20, 9, 4, 6, 6, 22, 5, 15, 21, 8, 18, 16, 22, 7, 8, 21, 9, 5, 19, 9, 22, 20, 23, 10, 15, 4, 15, 24, 17, 1, 14, 6, 21, 8, 5, 24, 16, 18, 12, 5, 13, 15, 15, 12, 6, 4, 3, 16, 13, 4, 4, 16, 24, 14, 9, 15, 21, 8, 5, 20, 21, 18, 5, 6, 21, 12, 9, 21, 21, 12, 5, 6, 5, 4, 9, 6, 20, 15, 6, 24, 10, 14, 4, 24, 6, 18, 5, 24, 9, 9, 18, 13, 10, 14, 7, 5, 22, 19, 20, 2, 15, 4, 20, 16, 19, 14, 16, 2, 17, 5, 18, 10, 15, 20, 15, 20, 17, 9, 18, 2, 13, 19, 1, 15, 5, 20, 8, 16, 22, 7, 8, 21, 9, 5, 19, 22, 15, 23, 1, 20, 20, 8, 9, 15, 10, 14, 7, 2, 15, 4, 20, 9, 6, 19, 11, 26, 2, 8, 1, 19, 20, 8, 2, 13, 22, 5, 20, 9, 6, 18, 5, 20, 6, 5, 13, 6, 5, 20, 15, 3, 6, 14, 15, 4, 16, 12, 15, 
+22, 19, 9, 14, 2, 15, 25, 20, 9, 10, 14, 7, 6, 25, 3, 5, 17, 21, 20, 8, 6, 17, 15, 19, 21, 6, 18, 19, 21, 9, 1, 20, 24, 6, 18, 5, 17, 13, 1, 19, 21, 6, 18, 5, 5, 6, 22, 5, 19, 26, 23, 8, 6, 19, 5, 20, 9, 6, 2, 12, 2, 4, 11, 13, 
+16, 22, 19, 20, 2, 4, 8, 9, 16, 5, 6, 1, 4, 6, 7, 1, 1, 6, 4, 4, 16, 24, 14, 6, 19, 16, 13, 5, 23, 6, 18, 25, 4, 16, 13, 13, 2, 15, 4, 9, 15, 8, 3, 15, 19, 15, 5, 18, 21, 9, 5, 18, 6, 24, 1, 19, 16, 15, 5, 15, 15, 21, 8, 5, 9, 16, 21, 19, 6, 7, 18, 15, 15, 21, 9, 13, 14, 6, 4, 9, 2, 21, 5, 12, 26, 16, 16, 16, 16, 20, 9, 20, 6, 3, 9, 7, 3, 19, 15, 20, 9, 6, 18, 9, 20, 24, 1, 20, 4, 9, 9, 14, 8, 26, 15, 21, 21, 9, 5, 3, 2, 17, 20, 9, 16, 15, 19, 1, 10, 5, 23, 8, 10, 13, 5, 20, 9, 6, 4, 1, 19, 12, 5, 25, 6, 20, 12, 15, 16, 12, 5, 4, 5, 6, 5, 16, 10, 15, 20, 15, 24, 10, 14, 19, 21, 16, 14, 19, 16, 24, 14, 4, 16, 24, 14, 1, 21, 20, 20, 18, 6, 6, 20, 12, 6, 23, 5, 12, 2, 15, 15, 20, 9, 6, 18, 16, 16, 20, 20, 5, 19, 21, 15, 18, 15, 2, 20, 15, 15, 6, 3, 15, 19, 15, 5, 18, 7, 13, 1, 16, 17, 6, 4, 6, 10, 21, 6, 21, 13, 13, 25, 9, 15, 21, 8, 5, 24, 10, 14, 4, 2, 13, 20, 5, 19, 15, 1, 20, 6, 13, 25, 3, 16, 23, 
+5, 18, 10, 15, 7, 1, 15, 5, 21, 14, 4, 16, 22, 5, 19, 10, 14, 7, 21, 9, 5, 19, 10, 15, 7, 12, 6, 24, 15, 18, 5, 10, 14, 7, 20, 16, 3, 9, 15, 21, 8, 5, 7, 2, 18, 4, 10, 20, 20, 1, 15, 4, 5, 1, 9, 6, 12, 9, 4, 16, 16, 20, 6, 19, 19, 11, 10, 14, 13, 5, 5, 5, 15, 23, 15, 3, 5, 20, 24, 6, 5, 14, 21, 9, 5, 18, 16, 16, 6, 19, 9, 16, 22, 5, 19, 6, 
+4, 6, 16, 19, 1, 14, 10, 15, 19, 20, 2, 15, 20, 12, 10, 12, 5, 1, 3, 13, 21, 5, 3, 16, 20, 20, 13, 6, 1, 14, 5, 5, 1, 18, 21, 6, 4, 1, 24, 2, 25, 1, 8, 2, 9, 14, 24, 10, 20, 8, 2, 4, 21, 18, 23, 10, 14, 7, 7, 13, 9, 7, 9, 21, 9, 20, 24, 2, 19, 20, 9, 6, 16, 15, 13, 10, 3, 5, 17, 2, 20, 18, 16, 13, 19, 14, 16, 16, 16, 9, 15, 8, 9, 14, 21, 16, 16, 5, 16, 17, 12, 5, 20, 24, 9, 14, 5, 16, 23, 19, 21, 9, 5, 16, 2, 21, 18, 15, 13, 20, 4, 9, 5, 15, 15, 20, 14, 2, 20, 20, 6, 19, 8, 15, 24, 6, 22, 5, 19, 16, 14, 12, 26, 21, 8, 5, 21, 9, 15, 21, 8, 9, 20, 16, 16, 13, 9, 3, 
+6, 14, 1, 20, 21, 6, 18, 5, 5]
+
+periodic_affine_123_765_orwell_tokens = [16, 12, 12, 8, 9, 6, 9, 6, 20, 14, 2, 23, 10, 23, 9, 11, 16, 6, 6, 5, 19, 8, 26, 13, 16, 14, 6, 21, 16, 23, 15, 19, 16, 19, 23, 16, 18, 9, 12, 12, 6, 26, 26, 12, 13, 16, 11, 20, 21, 25, 23, 15, 5, 13, 1, 19, 26, 21, 21, 20, 21, 9, 23, 22, 20, 18, 20, 5, 23, 15, 2, 20, 26, 13, 15, 16, 20, 19, 2, 4, 1, 19, 19, 21, 16, 20, 23, 22, 2, 20, 26, 
+10, 13, 12, 7, 18, 1, 5, 19, 8, 20, 26, 13, 22, 24, 25, 12, 23, 22, 19, 18, 10, 7, 3, 12, 12, 15, 12, 18, 20, 19, 
+19, 12, 16, 20, 21, 26, 14, 20, 23, 26, 26, 11, 3, 2, 16, 13, 4, 19, 1, 23, 15, 6, 24, 2, 25, 15, 1, 2, 26, 14, 14, 6, 26, 9, 21, 22, 23, 13, 26, 23, 5, 3, 5, 16, 1, 23, 13, 6, 17, 6, 21, 9, 20, 22, 20, 18, 1, 2, 24, 2, 25, 15, 
+21, 23, 23, 24, 15, 20, 10, 11, 9, 6, 19, 19, 22, 15, 10, 15, 12, 24, 23, 6, 26, 3, 19, 19, 1, 7, 18, 4, 5, 13, 19, 23, 5, 14, 6, 20, 1, 12, 22, 11, 15, 18, 1, 22, 13, 22, 17, 26, 21, 12, 26, 25, 5, 19, 14, 7, 9, 22, 20, 10, 4, 
+5, 23, 15, 2, 20, 20, 12, 15, 12, 2, 6, 19, 14, 12, 8, 1, 18, 20, 19, 9, 1, 23, 5, 9, 23, 20, 19, 19, 21, 10, 7, 11, 9, 7, 10, 12, 7, 19, 11, 23, 9, 11, 6, 6, 14, 17, 6, 1, 9, 6, 1, 23, 19, 12, 19, 19, 11, 23, 5, 16, 12, 6, 10, 
+23, 9, 22, 15, 13, 12, 16, 3, 22, 9, 23, 12, 6, 23, 22, 23, 9, 8, 6, 10, 12, 22, 24, 25, 5, 19, 11, 23, 24, 25, 16, 20, 26, 26, 9, 8, 1, 15, 8, 16, 11, 12, 19, 19, 1, 7, 16, 18, 19, 21, 1, 23, 23, 15, 19, 12, 8, 14, 9, 16, 12, 21, 12, 26, 20, 10, 12, 26, 11, 9, 20, 20, 26, 9, 6, 7, 19, 12, 20, 24, 25, 17, 24, 2, 9, 5, 8, 13, 26, 20, 23, 13, 12, 12, 15, 8, 20, 6, 20, 19, 23, 25, 19, 12, 16, 16, 26, 1, 2, 26, 13, 7, 16, 12, 23, 5, 8, 17, 6, 21, 23, 5, 8, 10, 24, 2, 12, 5, 22, 6, 23, 6, 22, 20, 3, 19, 12, 16, 12, 15, 8, 2, 26, 8, 18, 22, 9, 14, 6, 10, 11, 14, 22, 15, 18, 1, 7, 16, 15, 19, 6, 21, 16, 13, 2, 25, 10, 12, 16, 9, 6, 2, 6, 21, 16, 18, 22, 17, 26, 13, 19, 6, 1, 15, 13, 12, 9, 12, 16, 20, 18, 1, 23, 19, 20, 7, 21, 12, 22, 24, 25, 12, 15, 12, 9, 23, 8, 5, 13, 26, 5, 23, 4, 7, 18, 21, 23, 2, 26, 19, 23, 25, 1, 20, 21, 25, 23, 15, 19, 9, 16, 22, 23, 12, 18, 26, 21, 7, 23, 1, 2, 26, 9, 19, 18, 1, 
+23, 5, 1, 5, 14, 12, 9, 20, 1, 21, 6, 26, 9, 26, 19, 16, 24, 20, 21, 24, 25, 11, 20, 21, 25, 6, 21, 16, 6, 1, 26, 
+13, 12, 9, 26, 21, 12, 23, 15, 19, 26, 19, 19, 16, 1, 6, 20, 10, 13, 2, 25, 6, 26, 21, 12, 12, 8, 9, 16, 2, 12, 24, 13, 22, 21, 2, 6, 20, 21, 25, 21, 8, 1, 9, 16, 25, 15, 1, 2, 24, 2, 6, 18, 16, 12, 12, 8, 9, 3, 8, 6, 23, 22, 22, 23, 15, 19, 26, 10, 23, 19, 22, 17, 22, 11, 6, 20, 3, 19, 20, 21, 26, 13, 12, 26, 6, 25, 7, 23, 16, 23, 19, 13, 
+23, 13, 15, 7, 23, 12, 21, 26, 12, 11, 23, 15, 19, 5, 19, 7, 23, 4, 7, 18, 26, 19, 7, 12, 20, 5, 19, 5, 10, 15, 12, 18, 2, 26, 6, 21, 16, 12, 16, 20, 18, 1, 23, 19, 4, 2, 24, 4, 7, 18, 1, 2, 20, 25, 12, 22, 21, 5, 19, 12, 7, 19, 11, 2, 6, 11, 7, 7, 8, 6, 20, 10, 23, 18, 12, 15, 9, 10, 19, 13, 8, 10, 24, 3, 19, 15, 16, 9, 13, 16, 25, 15, 1, 
+7, 19, 18, 14, 26, 4, 19, 19, 1, 9, 9, 22, 21, 9, 6, 6, 26, 26, 12, 20, 21, 25, 18, 12, 18, 26, 25, 7, 9, 1, 5, 14, 12, 9, 24, 21, 12, 15, 12, 21, 6, 6, 23, 19, 12, 7, 16, 15, 14, 6, 21, 16, 20, 21, 25, 24, 23, 26, 24, 26, 5, 23, 12, 12, 15, 12, 14, 20, 13, 12, 18, 15, 7, 5, 1, 12, 15, 12, 26, 24, 26, 12, 26, 25, 21, 20, 1, 2, 23, 15, 19, 26, 21, 23, 13, 20, 23, 2, 26, 22, 6, 10, 19, 10, 8, 4, 26, 11, 22, 13, 22, 17, 23, 15, 19, 12, 8, 14, 9, 16, 12, 12, 8, 9, 24, 21, 19, 24, 13, 12, 15, 22, 9, 26, 23, 5, 16, 1, 15, 13, 12, 9, 12, 15, 5, 16, 15, 7, 13, 12, 9, 24, 
+10, 23, 19, 1, 6, 20, 3, 19, 21, 1, 2, 6, 1, 12, 15, 12, 19, 22, 12, 9, 5, 22, 14, 9, 22, 21, 22, 22, 15, 6, 9, 23, 2, 1, 21, 15, 12, 20, 22, 22, 15, 14, 22, 18, 26, 9, 5, 10, 9, 6, 24, 1, 2, 26, 25, 5, 18, 4, 7, 23, 10, 2, 20, 
+21, 25, 22, 22, 15, 23, 15, 19, 16, 8, 26, 23, 16, 23, 19, 9, 19, 19, 12, 7, 23, 15, 5, 23, 25, 7, 19, 16, 20, 18, 16, 16, 26, 1, 2, 26, 13, 14, 6, 1, 7, 5, 25, 15, 20, 1, 1, 7, 22, 5, 16, 12, 21, 6, 26, 6, 26, 8, 16, 20, 21, 25, 24, 2, 12, 6, 19, 5, 18, 1, 23, 5, 13, 5, 10, 2, 6, 26, 26, 21, 15, 16, 13, 15, 15, 7, 21, 26, 23, 14, 12, 12, 15, 16, 20, 10, 1, 23, 21, 22, 21, 20, 1, 2, 23, 15, 19, 3, 25, 23, 21, 2, 13, 23, 16, 23, 19, 22, 22, 3, 16, 25, 20, 25, 23, 19, 1, 2, 26, 3, 23, 20, 10, 19, 16, 8, 17, 26, 13, 6, 24, 20, 7, 19, 22, 10, 9, 22, 20, 10, 20, 19, 23, 8, 14, 3, 19, 7, 8, 2, 19, 9, 16, 11, 26, 8, 16, 2, 19, 14, 26, 11, 17, 20, 25, 6, 24, 25, 21, 15, 16, 13, 15, 13, 23, 13, 20, 19, 21, 23, 7, 13, 1, 23, 5, 1, 2, 26, 26, 15, 13, 13, 7, 16, 12, 23, 5, 1, 2, 26, 25, 5, 10, 15, 12, 15, 8, 20, 21, 4, 7, 9, 19, 21, 20, 21, 9, 23, 22, 20, 23, 2, 6, 19, 12, 16, 6, 26, 21, 20, 1, 13, 15, 8, 20, 21, 1, 2, 26, 3, 23, 20, 10, 19, 18, 8, 20, 4, 26, 23, 14, 12, 21, 15, 8, 12, 23, 15, 23, 2, 14, 2, 23, 15, 19, 12, 22, 6, 21, 26, 21, 26, 25, 19, 18, 1, 5, 9, 19, 16, 20, 26, 12, 20, 21, 25, 2, 16, 9, 15, 8, 10, 9, 12, 12, 15, 12, 5, 19, 26, 12, 13, 2, 17, 26, 21, 12, 23, 15, 19, 23, 12, 14, 26, 26, 13, 13, 12, 19, 19, 16, 12, 12, 8, 9, 16, 8, 14, 9, 12, 16, 16, 22, 15, 9, 11, 10, 26, 11, 5, 14, 20, 19, 21, 9, 15, 23, 1, 2, 26, 25, 19, 12, 8, 9, 19, 22, 21, 6, 6, 23, 5, 26, 2, 2, 1, 12, 20, 21, 25, 20, 1, 23, 5, 13, 13, 24, 20, 26, 9, 12, 12, 26, 19, 1, 15, 12, 17, 24, 3, 19, 21, 22, 18, 26, 25, 12, 24, 1, 2, 26, 4, 5, 19, 11, 23, 12, 8, 9, 14, 8, 14, 9, 16, 9, 15, 13, 6, 6, 
+16, 14, 5, 16, 25, 2, 25, 19, 23, 15, 19, 14, 12, 7, 10, 25, 19, 19, 12, 9, 18, 22, 22, 15, 16, 9, 11, 22, 16, 22, 20, 19, 13, 12, 14, 22, 12, 17, 3, 15, 7, 18, 16, 4, 26, 11, 10, 22, 1, 2, 26, 9, 14, 2, 12, 23, 7, 12, 6, 6, 19, 14, 18, 4, 2, 20, 10, 2, 12, 12, 6, 26, 1, 2, 26, 2, 20, 20, 13, 23, 13, 20, 23, 5, 1, 2, 26, 23, 7, 13, 1, 1, 15, 16, 9, 15, 8, 5, 13, 4, 7, 18, 3, 19, 13, 6, 22, 6, 16, 6, 15, 16, 9, 5, 8, 13, 26, 21, 7, 23, 2, 6, 6, 19, 14, 
+22, 26, 7, 19, 14, 15, 20, 21, 19, 15, 16, 9, 18, 18, 5, 19, 25, 23, 2, 14, 2, 26, 21, 19, 21, 9, 1, 16, 22, 7, 13, 26, 19, 18, 22, 7, 3, 8, 20, 21, 9, 14, 2, 21, 12, 13, 8, 4, 24, 25, 10, 9, 8, 16, 26, 26, 7, 19, 11, 12, 15, 12, 13, 24, 19, 16, 24, 13, 12, 15, 12, 21, 20, 21, 12, 26, 25, 12, 15, 8, 12, 15, 8, 16, 25, 2, 9, 23, 12, 20, 21, 
+12, 16, 24, 2, 12, 18, 16, 16, 26, 12, 18, 26, 21, 12, 15, 25, 23, 2, 14, 2, 23, 15, 19, 18, 15, 15, 23, 4, 5, 19, 11, 23, 12, 23, 7, 19, 12, 12, 15, 12, 21, 24, 25, 14, 21, 19, 23, 24, 18, 19, 21, 10, 23, 9, 11, 16, 24, 4, 20, 
+20, 21, 12, 15, 12, 9, 23, 25, 19, 26, 1, 14, 20, 1, 12, 9, 12, 19, 21, 11, 5, 26, 26, 23, 5, 4, 5, 19, 11, 21, 26, 25, 19, 12, 15, 5, 13, 19, 5, 19, 14, 16, 2, 26, 12, 6, 21, 16, 23, 22, 6, 19, 23, 7, 3, 12, 6, 20, 21, 12, 24, 
+26, 26, 20, 25, 7, 9, 26, 7, 19, 11, 12, 15, 22, 15, 10, 15, 12, 15, 12, 9, 2, 21, 21, 6, 26, 9, 15, 16, 20, 20, 21, 25, 6, 21, 16, 23, 15, 19, 18, 18, 1, 6, 15, 7, 13, 26, 2, 11, 19, 15, 26, 1, 2, 26, 25, 19, 18, 12, 19, 14, 12, 16, 23, 22, 10, 26, 21, 23, 16, 22, 14, 24, 2, 6, 20, 21, 7, 19, 6, 12, 15, 16, 20, 10, 12, 24, 16, 12, 26, 23, 
+1, 2, 26, 23, 23, 18, 1, 19, 13, 26, 12, 15, 8, 12, 12, 12, 6, 26, 23, 14, 6, 26, 12, 26, 25, 19, 21, 12, 18, 26, 
+25, 1, 12, 15, 19, 13, 12, 12, 15, 12, 10, 9, 8, 13, 4, 20, 23, 2, 26, 12, 6, 10, 2, 20, 22, 16, 5, 8, 13, 26, 14, 7, 1, 12, 16, 21, 22, 21, 19, 13, 6, 24, 20, 19, 7, 12, 6, 22, 10, 23, 14, 20, 7, 19, 11, 5, 19, 14, 13, 24, 25, 
+20, 26, 25, 12, 15, 12, 6, 26, 4, 7, 18, 22, 20, 26, 22, 20, 23, 15, 19, 15, 22, 15, 18, 12, 22, 13, 22, 20, 23, 16, 17, 14, 12, 16, 20, 8, 12, 26, 19, 1, 24, 23, 26, 24, 26, 5, 23, 12, 10, 20, 14, 10, 13, 22, 12, 15, 12, 6, 20, 26, 21, 6, 1, 13, 15, 16, 20, 10, 6, 23, 2, 1, 2, 26, 10, 7, 3, 1, 5, 24, 21, 9, 6, 16, 16, 12, 15, 5, 9, 12, 12, 15, 12, 16, 6, 25, 11, 26, 6, 19, 18, 19, 23, 24, 18, 19, 21, 11, 19, 26, 23, 5, 19, 1, 23, 12, 16, 20, 18, 1, 23, 19, 26, 23, 12, 21, 16, 24, 4, 20, 6, 1, 9, 23, 25, 19, 26, 1, 14, 26, 3, 19, 9, 8, 20, 24, 1, 2, 26, 25, 26, 24, 26, 12, 26, 25, 12, 24, 25, 20, 6, 1, 23, 19, 12, 13, 24, 25, 20, 26, 25, 22, 9, 8, 26, 3, 12, 16, 5, 16, 12, 5, 2, 14, 9, 6, 5, 19, 1, 2, 26, 4, 5, 19, 11, 7, 9, 1, 19, 13, 21, 7, 23, 12, 14, 22, 10, 23, 7, 12, 6, 20, 21, 25, 6, 21, 16, 2, 21, 13, 24, 3, 19, 13, 16, 20, 10, 1, 2, 26, 26, 5, 19, 14, 14, 26, 4, 23, 13, 11, 5, 19, 14, 9, 24, 10, 5, 19, 1, 2, 26, 13, 7, 13, 11, 5, 18, 1, 7, 19, 10, 19, 6, 15, 19, 9, 16, 13, 24, 23, 12, 26, 25, 9, 4, 16, 17, 14, 12, 16, 21, 22, 21, 19, 9, 19, 23, 4, 19, 26, 21, 12, 15, 12, 6, 24, 22, 22, 18, 15, 23, 7, 12, 6, 26, 11, 22, 24, 25, 7, 19, 16, 20, 18, 1, 7, 19, 1, 14, 20, 18, 19, 6, 9, 14, 2, 12, 10, 24, 1, 12, 9, 12, 7, 19, 11, 16, 6, 25, 12, 26, 11, 7, 12, 8, 1, 6, 14, 7, 20, 21, 21, 20, 1, 2, 6, 10, 15, 13, 3, 5, 19, 14, 22, 9, 16, 25, 15, 
+1, 5, 23, 4, 7, 18, 1, 2, 26, 23, 23, 9, 16, 13, 26, 23, 7, 23, 25, 23, 9, 26, 20, 24, 22, 26, 20, 21, 25, 20, 21, 12, 24, 23, 19, 24, 23, 14, 26, 26, 21, 20, 21, 16, 24, 4, 9, 23, 15, 19, 3, 8, 12, 13, 22, 14, 18, 11, 5, 21, 21, 23, 23, 20, 7, 23, 1, 19, 13, 15, 23, 12, 12, 18, 26, 25, 23, 19, 19, 1, 23, 15, 19, 23, 15, 23, 2, 14, 2, 23, 23, 23, 9, 16, 13, 26, 20, 7, 23, 1, 19, 13, 12, 16]
+
+columnar_1432_orwell_tokens = [9, 19, 9, 3, 4, 14, 9, 4, 3, 11, 18, 18, 14, 9, 5, 14, 14, 20, 19, 14, 26, 9, 8, 18, 20, 14, 15, 15, 1, 8, 12, 14, 9, 4, 3, 20, 21, 8, 1, 15, 15, 3, 25, 19, 19, 21, 15, 9, 25, 21, 15, 22, 1, 18, 7, 20, 19, 15, 20, 14, 15, 9, 9, 5, 12, 19, 20, 15, 4, 2, 1, 12, 7, 19, 14, 4, 20, 12, 5, 19, 20, 1, 6, 14, 18, 16, 8, 5, 1, 4, 8, 12, 4, 3, 19, 12, 5, 13, 6, 13, 20, 1, 18, 4, 5, 5, 13, 6, 21, 18, 9, 9, 8, 25, 3, 21, 3, 14, 7, 12, 14, 13, 1, 5, 14, 14, 5, 20, 20, 19, 1, 21, 18, 7, 12, 5, 1, 5, 20, 9, 9, 19, 4, 15, 14, 4, 18, 14, 5, 3, 3, 18, 23, 21, 6, 9, 1, 7, 15, 9, 19, 20, 8, 15, 25, 22, 16, 1, 9, 15, 20, 5, 5, 20, 19, 14, 7, 21, 4, 19, 23, 1, 9, 14, 1, 1, 1, 15, 12, 1, 5, 18, 20, 12, 14, 15, 18, 9, 5, 1, 13, 14, 23, 14, 8, 
+4, 15, 19, 20, 9, 8, 20, 15, 18, 8, 5, 13, 6, 7, 4, 13, 23, 9, 19, 15, 15, 9, 18, 8, 1, 15, 20, 5, 1, 5, 19, 12, 15, 15, 8, 15, 22, 7, 20, 9, 20, 
+14, 21, 3, 9, 5, 20, 18, 14, 5, 6, 1, 9, 15, 23, 5, 14, 20, 19, 6, 18, 8, 8, 15, 8, 20, 23, 20, 18, 3, 14, 9, 15, 5, 3, 13, 15, 15, 14, 20, 12, 5, 5, 12, 13, 15, 9, 15, 4, 20, 8, 18, 5, 8, 7, 1, 1, 9, 15, 18, 1, 20, 14, 5, 3, 14, 13, 1, 15, 20, 15, 23, 19, 12, 20, 21, 1, 20, 14, 21, 20, 20, 
+19, 5, 23, 1, 4, 12, 4, 5, 20, 18, 19, 1, 19, 20, 9, 6, 16, 5, 5, 5, 5, 20, 9, 23, 1, 19, 1, 9, 5, 13, 18, 19, 8, 15, 5, 25, 8, 26, 25, 2, 15, 1, 
+23, 8, 5, 21, 15, 6, 16, 25, 8, 23, 5, 1, 9, 3, 20, 12, 1, 9, 9, 9, 21, 14, 25, 18, 15, 14, 21, 1, 2, 5, 4, 3, 15, 5, 20, 8, 1, 19, 4, 21, 4, 5, 18, 8, 19, 23, 15, 14, 5, 12, 15, 3, 4, 9, 5, 5, 9, 5, 9, 6, 4, 5, 18, 7, 20, 20, 16, 18, 15, 18, 1, 8, 8, 19, 1, 9, 7, 20, 11, 1, 2, 20, 5, 13, 15, 15, 15, 14, 20, 7, 5, 8, 19, 19, 20, 5, 19, 5, 5, 8, 20, 12, 13, 20, 9, 1, 1, 4, 6, 5, 25, 13, 9, 15, 18, 18, 19, 15, 5, 19, 15, 13, 9, 12, 16, 
+20, 7, 20, 9, 20, 14, 21, 3, 9, 1, 8, 20, 1, 25, 15, 4, 16, 15, 19, 19, 4, 1, 18, 12, 12, 20, 16, 5, 18, 15, 15, 18, 16, 6, 21, 9, 5, 4, 5, 20, 3, 18, 1, 14, 5, 7, 19, 12, 18, 7, 9, 5, 4, 1, 1, 9, 20, 11, 5, 23, 20, 14, 18, 19, 5, 6, 14, 20, 12, 1, 5, 20, 14, 18, 1, 1, 14, 8, 18, 7, 7, 20, 20, 15, 5, 18, 14, 9, 14, 5, 5, 14, 19, 16, 15, 9, 20, 20, 15, 5, 12, 5, 21, 16, 3, 20, 5, 1, 18, 20, 4, 9, 18, 14, 5, 3, 5, 20, 9, 8, 5, 9, 15, 9, 9, 9, 26, 4, 15, 2, 19, 1, 6, 20, 3, 20, 9, 9, 12, 5, 9, 25, 15, 20, 12, 4, 19, 9, 18, 14, 14, 15, 14, 21, 12, 15, 20, 5, 20, 9, 6, 20, 21, 18, 14, 9, 12, 23, 8, 8, 12, 25, 12, 2, 5, 2, 5, 15, 1, 20, 15, 14, 9, 15, 18, 15, 18, 12, 5, 9, 15, 19, 25, 2, 20, 5, 20, 1, 20, 9, 4, 16, 14, 18, 19, 5, 5, 14, 20, 9, 8, 3, 1, 15, 15, 15, 6, 23, 1, 22, 1, 15, 1, 1, 21, 4, 1, 15, 5, 18, 9, 15, 4, 18, 19, 18, 23, 15, 20, 14, 5, 20, 14, 8, 19, 20, 19, 1, 12, 23, 9, 14, 16, 5, 8, 5, 9, 18, 20, 3, 6, 18, 4, 9, 8, 19, 1, 18, 20, 3, 13, 9, 14, 16, 20, 6, 1, 5, 8, 1, 19, 5, 9, 19, 14, 14, 14, 23, 8, 25, 5, 8, 22, 3, 21, 18, 22, 19, 8, 11, 5, 12, 25, 20, 19, 18, 9, 15, 5, 15, 3, 14, 7, 15, 5, 12, 19, 20, 16, 5, 20, 5, 18, 19, 5, 5, 15, 5, 
+12, 1, 5, 8, 16, 21, 23, 8, 19, 14, 22, 8, 8, 5, 12, 25, 2, 23, 25, 15, 9, 15, 18, 1, 9, 15, 5, 20, 2, 1, 20, 9, 4, 5, 20, 21, 22, 5, 18, 9, 21, 9, 6, 21, 23, 8, 19, 20, 7, 15, 8, 16, 21, 15, 16, 18, 8, 9, 1, 18, 14, 15, 5, 16, 21, 11, 21, 4, 18, 8, 6, 5, 18, 20, 21, 3, 20, 9, 8, 23, 23, 20, 21, 4, 9, 1, 8, 9, 1, 15, 8, 8, 8, 23, 19, 5, 12, 19, 7, 8, 5, 9, 18, 14, 5, 5, 5, 20, 3, 5, 21, 5, 13, 21, 5, 1, 23, 6, 20, 7, 6, 13, 20, 8, 22, 22, 15, 23, 15, 13, 9, 18, 6, 18, 5, 7, 5, 6, 2, 13, 12, 16, 9, 2, 5, 5, 18, 19, 3, 18, 5, 6, 15, 5, 20, 19, 18, 22, 6, 8, 1, 1, 1, 19, 21, 8, 11, 15, 5, 2, 1, 19, 1, 12, 18, 18, 4, 14, 5, 4, 8, 14, 20, 8, 21, 14, 15, 9, 22, 8, 7, 5, 20, 4, 1, 8, 18, 15, 4, 4, 14, 8, 18, 12, 12, 4, 15, 14, 
+18, 9, 14, 19, 4, 14, 5, 20, 9, 19, 20, 7, 5, 23, 8, 14, 4, 19, 8, 8, 5, 18, 5, 20, 14, 12, 9, 25, 14, 3, 20, 15, 18, 1, 18, 1, 18, 22, 23, 5, 2, 
+11, 19, 8, 6, 7, 4, 14, 13, 18, 13, 4, 3, 5, 5, 1, 5, 8, 21, 18, 9, 4, 5, 16, 9, 9, 15, 18, 1, 9, 15, 5, 20, 19, 23, 5, 4, 5, 12, 5, 5, 20, 14, 14, 14, 14, 20, 20, 5, 15, 18, 20, 15, 20, 3, 5, 1, 4, 6, 25, 8, 14, 20, 1, 25, 5, 7, 21, 22, 14, 5, 7, 15, 14, 3, 8, 18, 20, 5, 12, 16, 19, 13, 15, 5, 5, 5, 6, 22, 4, 1, 19, 20, 5, 21, 20, 1, 1, 4, 25, 9, 20, 21, 14, 9, 9, 19, 16, 3, 20, 19, 16, 9, 16, 12, 9, 23, 5, 18, 4, 15, 20, 8, 22, 14, 
+8, 15, 20, 9, 1, 18, 23, 2, 8, 12, 25, 16, 1, 8, 15, 23, 19, 11, 20, 20, 23, 20, 13, 8, 8, 21, 5, 20, 19, 1, 14, 6, 20, 19, 5, 22, 23, 19, 16, 21, 12, 18, 8, 7, 19, 18, 22, 15, 1, 15, 8, 8, 17, 11, 14, 8, 18, 14, 23, 15, 9, 4, 6, 5, 18, 1, 7, 8, 20, 1, 1, 5, 6, 12, 1, 7, 4, 18, 1, 20, 5, 6, 
+3, 21, 16, 5, 15, 7, 18, 15, 9, 1, 4, 14, 11, 15, 23, 9, 16, 5, 13, 1, 15, 21, 3, 18, 1, 5, 23, 20, 1, 6, 14, 2, 6, 25, 5, 8, 1, 12, 13, 20, 5, 18, 5, 8, 19, 6, 21, 23, 20, 1, 15, 5, 9, 20, 14, 5, 9, 8, 6, 5, 20, 5, 6, 5, 23, 5, 13, 11, 1, 20, 19, 20, 12, 18, 21, 14, 19, 15, 21, 7, 12, 20, 18, 23, 1, 6, 5, 15, 18, 9, 5, 1, 14, 8, 23, 20, 12, 1, 22, 12, 20, 1, 9, 15, 15, 20, 20, 14, 4, 1, 9, 5, 5, 15, 9, 7, 14, 23, 19, 12, 19, 7, 5, 20, 19, 8, 25, 1, 1, 14, 16, 20, 5, 20, 6, 5, 20, 9, 8, 15, 21, 3, 26, 18, 8, 12, 23, 14, 20, 5, 20, 19, 3, 5, 15, 9, 20, 20, 25, 15, 23, 1, 20, 14, 13, 2, 18, 5, 23, 8, 25, 8, 16, 14, 5, 9, 14, 9, 8, 1, 18, 25, 3, 19, 4, 15, 12, 15, 7, 19, 3, 4, 5, 14, 4, 20, 5, 4, 9, 6, 9, 20, 15, 3, 6, 1, 12, 13, 12, 17, 9, 4, 5, 18, 23, 8, 13, 1, 6, 19, 1, 6, 18, 20, 4, 12, 19, 20, 5, 23, 8, 20, 15, 19, 19, 23, 20, 7, 5, 4, 18, 9, 9, 14, 19, 12, 5, 
+20, 5, 8, 12, 18, 9, 19, 12, 15, 2, 13, 2, 8, 23, 15, 15, 21, 14, 15, 15, 5, 25, 15, 15, 20, 5, 4, 19, 12, 6, 12, 21, 8, 1, 14, 15, 19, 25, 5, 13, 19, 4, 8, 21, 5, 12, 9, 5, 8, 9, 13, 8, 18, 9, 9, 19, 25, 18, 6, 14, 18, 25, 7, 5, 19, 18, 8, 4, 15, 5, 16, 2, 20, 15, 1, 1, 8, 12, 20, 9, 18, 20, 10, 5, 4, 19, 5, 20, 21, 8, 21, 14, 16, 20, 15, 12, 5, 12, 23, 20, 20, 20, 20, 4, 19, 9, 5, 8, 9, 21, 14, 18, 16, 14, 16, 12, 4, 21, 8, 14, 19, 
+9, 14, 5, 1, 19, 21, 5, 5, 4, 5, 15, 18, 14, 9, 24, 20, 16, 5, 8, 5, 12, 5, 5, 25, 18, 5, 3, 21, 3, 4, 5, 5, 23, 15, 5, 15, 14, 7, 14, 8, 23, 14, 
+20, 15, 6, 20, 5, 20, 15, 19, 2, 18, 5, 23, 8, 25, 8, 16, 14, 4, 12, 5, 11, 19, 11, 5, 14, 9, 15, 23, 23, 19, 5, 22, 14, 5, 19, 20, 1, 5, 14, 12, 
+5, 20, 12, 20, 9, 12, 14, 12, 22, 14, 4, 15, 9, 8, 14, 23, 9, 15, 20, 1, 19, 3, 5, 15, 18, 13, 4, 2, 5, 8, 15, 15, 5, 18, 14, 14, 11, 12, 15, 5, 4, 5, 1, 1, 9, 3, 9, 12, 20, 1, 5, 9, 1, 12, 15, 7, 15, 16, 23, 15, 8, 20, 19, 14, 1, 18, 5, 15, 20, 8, 8, 12, 13, 5, 20, 1, 7, 15, 1, 1, 12, 20, 12, 19, 5, 9, 7, 18, 14, 19, 19, 8, 3, 14, 12, 14, 9, 5, 9, 5, 18, 5, 16, 5, 5, 4, 16, 17, 11, 8, 7, 5, 19, 15, 6, 20, 13, 9, 20, 7, 20, 3, 5, 7, 16, 5, 19, 12, 18, 25, 20, 13, 5, 7, 14, 20, 13, 8, 23, 13, 15, 9, 3, 1, 14, 4, 13, 1, 5, 15, 1, 15, 4, 20, 15, 18, 15, 4, 4, 12, 1, 5, 3, 20, 5, 12, 5, 20, 9, 25, 14, 15, 1, 15, 8, 13, 5, 5, 6, 15, 1, 1, 20, 20, 22, 20, 5, 2, 11, 19, 8, 4, 7, 25, 4, 5, 20, 19, 19, 13, 6, 8, 1, 9, 19, 19, 25, 20, 9, 22, 20, 2, 15, 13, 20, 19, 15, 18, 7, 1, 5, 20, 5, 20, 3, 5, 1, 20, 4, 14, 25, 8, 21, 20, 16, 15, 5, 14, 4, 5, 18, 18, 15, 18, 5, 11, 6, 23, 5, 6, 8, 16, 23, 20, 8, 19, 18, 9, 14, 4, 18, 19, 3, 2, 8, 9, 1, 5, 20, 23, 5, 14, 22, 12, 5, 20, 1, 5, 12, 9, 16, 9, 8, 6, 1, 8, 19, 23, 20, 14, 15, 1, 1, 6, 20, 1, 20, 15, 6, 19, 3, 5, 9, 18, 3, 18, 4, 20, 5, 6, 15, 21, 21, 5, 21, 5, 2, 8, 19, 3, 7, 20, 1, 15, 14, 8, 1, 19, 20, 12, 6, 20, 9, 1, 1, 7, 1, 20, 9, 5, 9, 1, 13, 9, 15, 9, 8, 15, 20, 15, 7, 14, 22, 5, 5, 13, 2, 7, 1, 1, 12, 1, 12, 9, 18, 3, 18, 16, 15, 5, 6, 15, 5, 8, 14, 12, 14, 14, 14, 19, 3, 4, 22, 5, 11, 5, 20, 21, 8, 18, 5, 20, 4, 9, 9, 2, 8, 19, 13, 20, 5, 3, 14, 1, 12, 3, 4, 9, 4, 20, 5, 14, 25, 8, 9, 20, 
+3, 12, 12, 13, 4, 18, 8, 14, 1, 12, 8, 9, 7, 20, 5, 5, 19, 9, 4, 18, 5, 1, 5, 20, 12, 22, 12, 8, 23, 20, 14, 18, 20, 1, 8, 1, 1, 18, 9, 19, 5, 21, 12, 14, 14, 19, 14, 7, 5, 3, 19, 1, 4, 14, 26, 12, 19, 20, 15, 6, 23, 5, 1, 4, 20, 5, 20, 5, 14, 15, 20, 8, 9, 23, 5, 23, 4, 11, 15, 15, 14, 19, 
+5, 20, 5, 5, 23, 23, 23, 12, 4, 1, 15, 1, 9, 19, 1, 14, 15, 20, 21, 19, 14, 1, 8, 25, 18, 12, 8, 19, 5, 2, 3, 21, 1, 8, 5, 16, 5, 20, 20, 23, 16, 
+20, 4, 18, 5, 8, 1, 15, 1, 15, 3, 26, 15, 18, 22, 3, 1, 14, 18, 20, 5, 15, 14, 8, 5, 14, 13, 1, 25, 15, 5, 2, 8, 19, 3, 7, 20, 1, 15, 9, 9, 8, 18, 5, 15, 4, 9, 23, 20, 15, 15, 20, 5, 5, 1, 8, 15, 18, 14, 14, 18, 6, 16, 9, 12, 14, 23, 1, 18, 5, 15, 9, 14, 3, 18, 20, 9, 5, 4, 19, 14, 6, 9, 14, 8, 3, 5, 9, 4, 14, 23, 20, 15, 8, 18, 15, 9, 1, 9, 2, 2, 12, 4, 20, 23, 7, 23, 1, 22, 6, 8, 23, 8, 12, 16, 15, 15, 14, 20, 15, 19, 4, 20, 1, 12, 
+4, 13, 5, 23, 18, 25, 20, 7, 15, 5, 20, 4]
+
+permutation_15324_orwell_tokens = [9, 19, 23, 20, 1, 1, 7, 18, 2, 9, 8, 12, 3, 20, 15, 4, 9, 1, 4, 25, 14, 9, 16, 1, 18, 12, 20, 14, 1, 4, 8, 15, 3, 5, 12, 3, 5, 19, 11, 23, 18, 18, 19, 5, 20, 9, 7, 9, 11, 14, 20, 20, 9, 8, 18, 5, 9, 14, 5, 23, 14, 14, 20, 19, 15, 19, 8, 9, 13, 20, 8, 8, 19, 9, 3, 9, 26, 14, 14, 21, 26, 9, 5, 12, 4, 14, 9, 15, 20, 8, 19, 1, 18, 2, 5, 19, 1, 9, 20, 14, 14, 15, 6, 5, 6, 18, 5, 20, 20, 15, 19, 5, 1, 3, 16, 20, 9, 5, 8, 22, 12, 14, 23, 5, 9, 4, 16, 12, 19, 9, 16, 21, 4, 5, 17, 9, 25, 11, 3, 12, 20, 21, 18, 8, 15, 7, 5, 20, 8, 8, 7, 19, 1, 12, 19, 4, 19, 15, 15, 18, 15, 3, 22, 6, 9, 20, 13, 18, 15, 25, 1, 15, 19, 14, 9, 14, 15, 20, 19, 8, 21, 15, 8, 7, 14, 20, 3, 21, 17, 9, 11, 14, 25, 12, 5, 15, 20, 7, 21, 8, 15, 22, 18, 16, 5, 5, 19, 20, 14, 1, 23, 15, 18, 9, 12, 6, 20, 18, 7, 9, 20, 19, 4, 25, 21, 20, 13, 18, 6, 15, 5, 18, 20, 14, 5, 9, 12, 7, 14, 1, 15, 9, 7, 14, 23, 20, 13, 8, 8, 9, 20, 1, 5, 8, 8, 12, 25, 23, 12, 1, 19, 20, 5, 13, 12, 15, 9, 2, 6, 15, 12, 1, 4, 5, 3, 2, 5, 1, 2, 7, 1, 12, 4, 14, 15, 4, 13, 1, 18, 7, 1, 20, 19, 20, 1, 15, 14, 5, 14, 5, 4, 20, 6, 15, 9, 1, 15, 15, 3, 12, 21, 16, 5, 18, 4, 15, 18, 20, 19, 5, 20, 1, 15, 15, 12, 18, 15, 
+5, 7, 6, 18, 15, 14, 9, 4, 15, 19, 4, 18, 9, 16, 8, 1, 12, 25, 1, 5, 2, 4, 5, 14, 11, 1, 20, 3, 5, 20, 20, 4, 15, 8, 12, 23, 5, 1, 12, 5, 20, 9, 4, 16, 5, 3, 9, 20, 4, 16, 9, 19, 13, 12, 5, 1, 25, 14, 14, 15, 18, 15, 13, 21, 3, 6, 19, 1, 5, 5, 15, 13, 18, 20, 1, 1, 8, 14, 13, 5, 20, 5, 18, 23, 20, 4, 9, 5, 8, 3, 6, 5, 1, 5, 13, 6, 15, 1, 1, 1, 15, 14, 6, 2, 6, 21, 15, 20, 15, 6, 20, 18, 25, 9, 9, 5, 22, 23, 20, 5, 1, 8, 8, 1, 12, 25, 
+22, 2, 1, 15, 11, 3, 13, 21, 3, 20, 19, 1, 8, 4, 1, 5, 14, 18, 5, 7, 21, 7, 4, 1, 25, 12, 8, 14, 13, 19, 4, 15, 5, 20, 5, 6, 1, 21, 23, 5, 18, 19, 9, 15, 19, 14, 20, 14, 5, 1, 13, 4, 6, 8, 18, 15, 20, 5, 9, 20, 19, 1, 18, 23, 9, 19, 20, 1, 21, 14, 19, 15, 19, 25, 20, 5, 18, 9, 8, 7, 14, 20, 
+5, 20, 9, 12, 6, 5, 1, 5, 22, 14, 20, 2, 8, 20, 5, 5, 6, 20, 19, 15, 20, 19, 13, 9, 5, 9, 19, 23, 20, 1, 19, 15, 12, 5, 4, 13, 11, 15, 23, 18, 9, 
+14, 7, 14, 1, 4, 18, 20, 1, 16, 5, 20, 5, 19, 14, 20, 12, 5, 8, 5, 5, 9, 20, 3, 18, 3, 18, 21, 3, 18, 5, 1, 20, 14, 23, 19, 15, 21, 3, 20, 6, 18, 
+4, 6, 21, 9, 1, 7, 14, 4, 25, 8, 9, 12, 7, 20, 18, 15, 8, 21, 19, 1, 20, 9, 23, 19, 20, 1, 16, 18, 15, 5, 20, 6, 8, 5, 15, 15, 3, 14, 13, 9, 4, 25, 18, 22, 16, 9, 5, 14, 18, 18, 16, 5, 1, 1, 14, 9, 20, 15, 6, 1, 18, 15, 8, 20, 5, 23, 5, 5, 11, 6, 8, 20, 5, 12, 1, 20, 1, 23, 19, 5, 5, 19, 22, 14, 7, 12, 6, 9, 8, 16, 19, 20, 21, 1, 9, 4, 14, 23, 14, 14, 20, 19, 15, 23, 1, 15, 8, 23, 19, 18, 8, 20, 9, 20, 14, 14, 25, 9, 5, 8, 14, 1, 4, 1, 1, 1, 4, 22, 18, 19, 3, 9, 15, 5, 5, 12, 21, 3, 18, 22, 2, 1, 15, 5, 18, 9, 8, 19, 9, 1, 8, 7, 20, 14, 23, 12, 11, 5, 5, 12, 20, 14, 19, 15, 18, 12, 23, 25, 5, 14, 20, 19, 9, 7, 5, 5, 19, 22, 18, 9, 12, 1, 20, 13, 14, 19, 5, 15, 20, 1, 5, 8, 23, 25, 1, 14, 15, 5, 3, 14, 12, 8, 1, 4, 15, 14, 9, 7, 16, 9, 15, 16, 19, 20, 5, 20, 5, 8, 12, 19, 6, 9, 20, 8, 20, 6, 1, 20, 8, 19, 16, 5, 15, 20, 9, 18, 5, 23, 20, 5, 20, 8, 8, 5, 13, 15, 14, 18, 15, 1, 19, 21, 6, 3, 26, 7, 5, 1, 5, 15, 6, 4, 18, 13, 23, 8, 20, 5, 1, 20, 12, 12, 9, 23, 14, 19, 1, 15, 5, 8, 6, 15, 20, 15, 9, 5, 19, 16, 
+3, 5, 21, 20, 18, 19, 3, 8, 23, 9, 8, 19, 18, 1, 5, 15, 20, 15, 3, 14, 18, 4, 22, 9, 5, 20, 20, 1, 8, 20, 8, 5, 5, 5, 25, 19, 12, 15, 6, 12, 15, 21, 25, 23, 15, 1, 20, 15, 2, 21, 23, 25, 5, 8, 14, 15, 22, 13, 21, 15, 5, 2, 9, 2, 7, 18, 5, 20, 15, 8, 18, 1, 19, 9, 23, 20, 14, 8, 3, 9, 7, 20, 
+15, 25, 21, 8, 16, 3, 5, 1, 20, 2, 15, 9, 14, 5, 20, 5, 14, 1, 8, 1, 20, 9, 18, 14, 9, 14, 9, 19, 4, 5, 20, 5, 8, 6, 1, 1, 12, 20, 6, 20, 21, 18, 
+9, 25, 3, 15, 22, 9, 5, 18, 1, 23, 19, 5, 14, 4, 1, 9, 7, 1, 21, 15, 20, 12, 15, 19, 9, 20, 6, 21, 9, 6, 7, 18, 8, 19, 5, 23, 9, 1, 8, 3, 8, 4, 5, 15, 19, 13, 20, 7, 9, 8, 14, 20, 23, 4, 15, 15, 9, 8, 8, 20, 20, 5, 4, 18, 16, 15, 21, 15, 20, 3, 9, 14, 9, 6, 15, 16, 7, 14, 18, 9, 15, 20, 15, 
+5, 8, 22, 9, 1, 5, 3, 3, 13, 15, 6, 5, 18, 13, 2, 14, 1, 15, 12, 13, 14, 15, 7, 5, 16, 1, 20, 12, 12, 5, 17, 1, 21, 12, 1, 11, 9, 5, 4, 5, 12, 21, 12, 4, 18, 9, 13, 18, 15, 9, 23, 18, 8, 3, 18, 6, 8, 15, 13, 1, 4, 5, 16, 18, 20, 15, 20, 6, 8, 18, 19, 5, 21, 6, 15, 3, 1, 5, 6, 18, 8, 20, 5, 9, 8, 8, 7, 20, 1, 1, 4, 14, 23, 12, 14, 23, 12, 9, 19, 20, 15, 20, 14, 21, 4, 14, 18, 5, 1, 20, 23, 19, 9, 3, 4, 1, 8, 14, 20, 15, 5, 8, 22, 9, 1, 5, 3, 19, 14, 13, 19, 11, 15, 5, 20, 8, 23, 1, 20, 7, 15, 8, 21, 8, 23, 8, 20, 5, 15, 23, 4, 18, 19, 5, 20, 5, 18, 19, 9, 9, 12, 12, 4, 19, 7, 9, 20, 14, 21, 1, 19, 9, 8, 2, 8, 5, 12, 20, 5, 20, 14, 9, 19, 18, 14, 13, 21, 5, 20, 20, 8, 20, 5, 5, 3, 5, 12, 19, 18, 9, 5, 5, 14, 20, 3, 1, 23, 
+19, 1, 4, 12, 12, 5, 3, 4, 21, 15, 12, 2, 13, 4, 5, 9, 13, 21, 4, 5, 2, 20, 18, 8, 20, 5, 5, 14, 1, 23, 19, 15, 15, 1, 23, 25, 6, 20, 8, 19, 21, 20, 9, 14, 9, 7, 20, 3, 6, 15, 6, 15, 5, 16, 13, 12, 20, 8, 12, 5, 25, 5, 5, 15, 13, 22, 4, 18, 22, 15, 5, 20, 5, 20, 15, 8, 23, 15, 14, 9, 4, 23, 
+1, 19, 1, 13, 12, 8, 9, 12, 19, 6, 12, 1, 18, 9, 6, 18, 7, 9, 21, 5, 13, 8, 20, 5, 5, 5, 7, 1, 18, 14, 15, 19, 5, 19, 6, 2, 9, 8, 19, 15, 5, 25, 4, 13, 18, 5, 12, 5, 25, 13, 19, 8, 16, 1, 9, 2, 5, 26, 4, 25, 2, 8, 20, 5, 12, 22, 5, 21, 15, 5, 12, 1, 18, 12, 19, 3, 8, 23, 9, 8, 5, 5, 23, 18, 
+20, 14, 5, 8, 21, 9, 13, 15, 6, 18, 15, 5, 20, 6, 8, 16, 25, 18, 1, 20, 8, 1, 19, 9, 8, 9, 19, 23, 18, 1, 22, 6, 18, 5, 25, 1, 9, 18, 9, 8, 19, 5, 1, 6, 3, 14, 18, 20, 1, 21, 1, 19, 12, 12, 25, 1, 9, 7, 14, 21, 14, 19, 8, 5, 9, 19, 18, 9, 11, 14, 15, 5, 7, 21, 8, 14, 25, 4, 5, 2, 3, 19, 1, 15, 18, 5, 16, 15, 19, 1, 1, 12, 4, 14, 2, 21, 1, 20, 14, 18, 26, 12, 18, 15, 2, 1, 1, 5, 4, 19, 14, 5, 20, 4, 8, 3, 15, 12, 15, 4, 6, 23, 8, 20, 5, 9, 18, 20, 14, 5, 20, 8, 1, 8, 20, 1, 19, 10, 4, 21, 20, 5, 14, 5, 4, 4, 19, 21, 15, 20, 9, 22, 5, 4, 5, 5, 18, 20, 14, 8, 15, 20, 7, 21, 8, 8, 
+21, 19, 5, 8, 20, 4, 9, 23, 14, 15, 14, 16, 23, 1, 5, 23, 8, 20, 5, 15, 12, 12, 18, 4, 15, 4, 11, 15, 5, 3, 4, 12, 15, 4, 15, 14, 14, 23, 9, 20, 20, 5, 8, 19, 18, 12, 5, 5, 20, 9, 5, 20, 20, 12, 5, 5, 4, 4, 9, 19, 9, 6, 15, 23, 14, 18, 23, 4, 5, 5, 18, 8, 23, 9, 12, 4, 14, 9, 7, 21, 14, 20, 
+19, 1, 4, 14, 15, 20, 18, 16, 18, 16, 1, 5, 9, 19, 20, 14, 15, 16, 12, 18, 9, 1, 19, 20, 14, 1, 4, 8, 8, 21, 15, 7, 20, 21, 5, 8, 19, 14, 19, 1, 23, 19, 8, 14, 14, 9, 9, 7, 20, 14, 1, 4, 8, 25, 19, 5, 11, 1, 19, 1, 8, 18, 8, 5, 12, 2, 21, 20, 5, 5, 8, 18, 19, 5, 5, 5, 13, 4, 5, 15, 20, 2, 14, 12, 3, 15, 15, 15, 14, 18, 21, 9, 1, 8, 25, 14, 20, 9, 24, 7, 14, 5, 3, 20, 16, 5, 20, 8, 19, 16, 5, 15, 20, 20, 18, 5, 19, 8, 5, 20, 1, 23, 18, 1, 16, 5, 12, 19, 5, 5, 20, 18, 4, 18, 22, 5, 5, 25, 18, 8, 23, 5, 5, 2, 8, 20, 5, 12, 13, 3, 1, 11, 15, 1, 19, 21, 20, 3, 4, 9, 8, 15, 6, 7, 3, 
+1, 5, 1, 4, 5, 26, 4, 15, 18, 14, 23, 6, 15, 5, 5, 13, 22, 18, 13, 3, 25, 15, 13, 9, 14, 1, 4, 14, 18, 3, 7, 15, 14, 8, 18, 5, 20, 5, 1, 5, 18, 23, 19, 15, 14, 15, 5, 14, 8, 8, 20, 5, 15, 6, 19, 21, 5, 18, 9, 14, 15, 20, 13, 9, 5, 13, 4, 1, 25, 5, 20, 12, 15, 19, 16, 16, 15, 9, 9, 5, 20, 2, 
+7, 20, 18, 2, 15, 8, 19, 18, 5, 9, 23, 8, 20, 1, 3, 9, 15, 7, 14, 25, 21, 3, 8, 20, 5, 1, 15, 20, 16, 9, 14, 4, 1, 19, 9, 23, 5, 9, 8, 12, 20, 1, 
+5, 8, 4, 18, 5, 5, 11, 25, 19, 11, 15, 12, 15, 5, 5, 4, 4, 5, 16, 15, 14, 9, 20, 23, 20, 14, 9, 19, 15, 23, 19, 14, 15, 14, 14, 15, 4, 23, 1, 18, 
+19, 20, 20, 5, 5, 20, 5, 12, 22, 14, 12, 5, 1, 15, 18, 8, 20, 5, 16, 5, 19, 15, 20, 18, 14, 15, 20, 18, 1, 5, 15, 20, 14, 3, 5, 18, 15, 14, 18, 16, 12, 6, 1, 16, 9, 4, 5, 6, 20, 12, 21, 6, 12, 25, 8, 14, 9, 20, 5, 4, 9, 23, 14, 1, 18, 20, 12, 5, 14, 12, 20, 1, 5, 25, 5, 15, 3, 22, 18, 1, 14, 9, 7, 14, 3, 21, 4, 14, 15, 9, 5, 22, 18, 14, 5, 20, 7, 8, 19, 12, 14, 9, 7, 5, 4, 15, 23, 18, 9, 15, 7, 14, 19, 3, 8, 14, 9, 20, 5, 4, 1, 6, 18, 9, 14, 20, 19, 1, 3, 5, 1, 5, 8, 12, 16, 3, 9, 15, 20, 11, 18, 5, 19, 9, 4, 13, 13, 5, 4, 2, 23, 15, 14, 5, 5, 23, 20, 5, 14, 18, 8, 20, 5, 15, 8, 6, 15, 19, 15, 5, 5, 22, 18, 4, 1, 15, 6, 18, 14, 20, 14, 9, 19, 1, 9, 20, 14, 12, 11, 12, 1, 5, 2, 21, 20, 2, 5, 15, 20, 14, 5, 12, 1, 4, 20, 1, 4, 18, 5, 1, 1, 4, 23, 25, 9, 7, 1, 1, 14, 8, 9, 23, 20, 1, 22, 21, 3, 18, 9, 12, 7, 14, 6, 9, 9, 8, 7, 20, 20, 20, 1, 23, 19, 8, 12, 16, 5, 15, 9, 1, 5, 3, 16, 20, 19, 15, 18, 12, 14, 9, 15, 15, 16, 14, 20, 9, 7, 14, 15, 16, 5, 16, 15, 12, 9, 19, 5, 23, 14, 19, 15, 4, 23, 20, 1, 5, 8, 16, 20, 19, 15, 18, 12, 4, 15, 4, 9, 14, 20, 20, 1, 13, 20, 5, 23, 8, 18, 15, 5, 15, 5, 22, 18, 14, 8, 25, 12, 20, 5, 21, 8, 20, 15, 7, 15, 20, 8, 16, 12, 13, 3, 9, 5, 1, 18, 20, 20, 5, 5, 4]
