--- sub_bench_reinforcement_10_u_05_fa.csv
@@ -1,151 +1,0 @@
--5.802347215335416
--5.639762485161516
--5.489591350179586
--5.359815739710143
--5.2347045143798745
--5.099369067426579
--4.985078879752137
--4.879036938120002
--4.77932237641385
--4.670183879541682
--4.601984761925537
--4.538618910745685
--4.500589746764643
--4.459236929345413
--4.393811106545687
--4.364718880865647
--4.310209106747681
--4.282043066996386
--4.264816216402801
--4.250240824779203
--4.232290579534559
--4.223857352645432
--4.221255350395918
--4.2211052624743965
--4.209526667660268
--4.208840134491695
--4.2025197256745495
--4.19526583192902
--4.189546280858029
--4.189546280858029
--4.187311547334581
--4.187311547334581
--4.187088522139672
--4.186865496944762
--4.185007651013271
--4.185007651013271
--4.185007651013271
--4.185007651013271
--4.185007651013271
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
