--- sub_bench_reinforcement_10_u_100_fa.csv
@@ -1,151 +1,0 @@
--5.805698669063272
--5.647908584433795
--5.485334387670309
--5.332201595074312
--5.169211985774516
--5.02968401280957
--4.874286650685534
--4.736163370288294
--4.603396932390307
--4.475786891260339
--4.3801380596885675
--4.3011295808735905
--4.2418763541538596
--4.211406067907049
--4.187787049457423
--4.1835001663231814
--4.182879140985455
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
