--- sub_bench_reinforcement_10_u_10_fa.csv
@@ -1,26 +1,0 @@
--5.804815412895934
--5.640417807927052
--5.50275867618825
--5.367792296473369
--5.2272115418915135
--5.075255812910303
--4.914761928206156
--4.765094901756935
--4.608245194974642
--4.505129295732481
--4.408754614047157
--4.321053433000186
--4.261502419572555
--4.221845392887227
--4.204885810416062
--4.192424292886734
--4.184804569291951
--4.183089775753828
--4.182826029237648
--4.1827729174898325
--4.1827729174898325
--4.1827729174898325
--4.1827729174898325
--4.1827729174898325
--4.1827729174898325
--4.1827729174898325
