--- sub_bench_reinforcement_10_u_20_fa.csv
@@ -1,151 +1,0 @@
--5.806768657421481
--5.631747556742649
--5.489055656779743
--5.344078869070332
--5.216018060860356
--5.052196875998832
--4.911795590037316
--4.743485977569167
--4.61141091638007
--4.4945780217465625
--4.391518457221399
--4.309108941823622
--4.245163559504613
--4.21273929501282
--4.203095597809908
--4.194879949297668
--4.183331736544418
--4.18302815327285
--4.182907176772944
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
