--- sub_bench_reinforcement_10_u_3_fa.csv
@@ -1,151 +1,0 @@
--5.805162345830135
--5.657295871859757
--5.525690499479633
--5.3859166482581875
--5.2372110466309785
--5.096352453309851
--4.96077232073531
--4.829488914055455
--4.725240132826348
--4.608304362077985
--4.5246503509745715
--4.445865808172376
--4.371363675482533
--4.320532034976624
--4.279361049030429
--4.243426986516212
--4.226870756319077
--4.212941284722489
--4.201135571449865
--4.1972986281252895
--4.190711421599156
--4.190513177257339
--4.1901806736324
--4.1879459401089525
--4.1879459401089525
--4.1879459401089525
--4.1879459401089525
--4.185711206585505
--4.185711206585505
--4.185711206585505
--4.185711206585505
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
--4.182772917489824
